% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub4_nmf.R
\name{assign_groups_nmf}
\alias{assign_groups_nmf}
\title{Assign Groups NMF}
\usage{
assign_groups_nmf(best_estimate, data)
}
\arguments{
\item{best_estimate}{Best estimate.}

\item{data}{Data.}
}
\description{
Assign Groups NMF
}
