% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_08_clinical_analysis.R
\name{AttachClinicalData}
\alias{AttachClinicalData}
\title{Attach Clinical Data and External Validation Set to Train_Model Object}
\usage{
AttachClinicalData(
  object,
  clinical_data = NULL,
  external_data = NULL,
  external_clinical = NULL
)
}
\arguments{
\item{object}{A Train_Model object.}

\item{clinical_data}{A data frame with clinical variables (rownames must match sample IDs).}

\item{external_data}{Optional matrix/data.frame for external validation dataset (expression/feature matrix).}

\item{external_clinical}{Optional data frame for external validation dataset clinical metadata.}
}
\value{
The updated Train_Model object.
}
\description{
Attach Clinical Data and External Validation Set to Train_Model Object
}
