% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-8-balance.R
\name{balance_and_visualize_data}
\alias{balance_and_visualize_data}
\title{Balance and visualize data distribution before and after balancing.}
\usage{
balance_and_visualize_data(
  data,
  group_col,
  method = "both",
  N = NULL,
  K = 5,
  dup_size = 0,
  seed = 123,
  palette_name = "Royal1",
  imbalance_threshold = 0.15,
  sample_size_threshold = 1500,
  force_balance = FALSE,
  save_plots = TRUE,
  save_dir = here("ModelData", "balacing_info"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)
}
\arguments{
\item{data}{A data frame containing the data to be balanced.}

\item{group_col}{A string specifying the column name that represents the class labels.}

\item{method}{A string specifying the balancing method. Can be "over", "under", or "both". Default is "both".}

\item{N}{A numeric value specifying the number of samples to generate after balancing. If \code{NULL}, it is automatically calculated based on the chosen method.}

\item{seed}{An integer to set the random seed for reproducibility. Default is 123.}

\item{palette_name}{A string specifying the color palette name for the plot. Default is "Royal1".}

\item{imbalance_threshold}{A numeric value representing the threshold for class imbalance. If the class ratio is below this threshold, balancing is triggered. Default is 0.15.}

\item{sample_size_threshold}{A numeric value representing the threshold for sample size. If the number of samples is below this threshold, balancing is triggered. Default is 1500.}

\item{force_balance}{A logical value indicating whether to force balancing regardless of the class distribution. Default is \code{FALSE}.}

\item{save_plots}{A logical value indicating whether to save the plot as a PDF. Default is \code{TRUE}.}

\item{save_dir}{A string specifying the directory where the plot will be saved. Default is "ModelData/balacing_info" using \code{here()}.}

\item{plot_width}{A numeric value specifying the width of the plot when saved. Default is 5.}

\item{plot_height}{A numeric value specifying the height of the plot when saved. Default is 5.}

\item{base_size}{An integer specifying the base font size for the plot. Default is 14.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{original_data}: The original (imbalanced) data.
\item \code{balanced_data}: The balanced dataset.
\item \code{method}: The method used for balancing.
\item \code{balanced_plot}: The plot visualizing the class distribution before and after balancing.
}
}
\description{
This function detects class imbalance in the dataset and balances it if necessary using specified methods. It also visualizes the class distribution before and after balancing, and optionally saves the plot as a PDF.
}
\examples{
result <- balance_and_visualize_data(data = dataset, group_col = "group")
result <- balance_and_visualize_data(data = dataset, group_col = "group", force_balance = TRUE)
}
