% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-3.R
\name{balance_data_process}
\alias{balance_data_process}
\title{Balance the dataset using different sampling methods.}
\usage{
balance_data_process(
  data,
  group_col = "group",
  method = "both",
  N = NULL,
  seed = 123
)
}
\arguments{
\item{data}{A data frame containing the dataset to be balanced.}

\item{group_col}{A string specifying the column name in the data frame that indicates the group labels (e.g., "group"). Default is "group".}

\item{method}{A string specifying the sampling method to use. It can be one of the following:
\itemize{
\item "over": Over-sampling the minority class.
\item "under": Under-sampling the majority class.
\item "both": Combination of both over-sampling and under-sampling. Default is "both".
}}

\item{N}{An integer specifying the desired number of rows in the balanced dataset. If \code{NULL}, it is automatically calculated based on the chosen method. Default is \code{NULL}.}

\item{seed}{An integer for random number generation to ensure reproducibility of the results. Default is 123.}
}
\value{
A data frame with balanced data.
}
\description{
Balance the dataset using different sampling methods.
}
\examples{
balanced_data <- balance_data_process(data = your_data, group_col = "group", method = "over", N = 1000)
balanced_data <- balance_data_process(data = your_data, method = "under")
}
