% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_08_normalize.R
\name{boxcox_transform}
\alias{boxcox_transform}
\title{Box-Cox Transformation}
\usage{
boxcox_transform(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Transformed vector.
}
\description{
Box-Cox Transformation
}
