% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_04_Extensions.R
\name{calibration_in_the_large}
\alias{calibration_in_the_large}
\title{Assess overall calibration: mean predicted vs observed survival probability}
\usage{
calibration_in_the_large(learner, object, time_point = NULL)
}
\arguments{
\item{learner}{A trained mlr3 learner (supports "distr").}

\item{object}{A `TaskSurv` or `PrognosiX` object.}

\item{time_point}{Time point for evaluation (default median event time).}
}
\value{
A list with calibration slope, intercept, and a plot.
}
\description{
Assess overall calibration: mean predicted vs observed survival probability
}
