% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_07_evaluate.R
\name{calinski_harabasz}
\alias{calinski_harabasz}
\title{Calinski-Harabasz Index}
\usage{
calinski_harabasz(clustered.data, label_col = "group")
}
\arguments{
\item{clustered.data}{Clustered data.}

\item{label_col}{label column.}
}
\value{
Numeric CH index value.
}
\description{
Calinski-Harabasz Index
}
\examples{
\dontrun{
  df <- data.frame(x = rnorm(30), y = rnorm(30), group = sample(1:3, 30, replace = TRUE))
  ch <- calinski_harabasz(df, label_col = "group")
  cat("Calinski-Harabasz Index:", ch, "\n")
}
}
