% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_05_train.R
\name{check_factor_level}
\alias{check_factor_level}
\title{Check the Number of Levels in the Factor Variable}
\usage{
check_factor_level(data, group_col)
}
\arguments{
\item{data}{A data frame or tibble containing the dataset.}

\item{group_col}{The name of the column to check.}
}
\value{
TRUE if the column contains at least two levels, otherwise stops.
}
\description{
Check the Number of Levels in the Factor Variable
}
