% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_02_FeatureSelect.R
\name{check_varImp_availability}
\alias{check_varImp_availability}
\title{Check Built-in Variable Importance Availability for caret Models}
\usage{
check_varImp_availability(model_names)
}
\arguments{
\item{model_names}{Character vector of caret model names.}
}
\value{
A data frame with model name and whether varImp is supported.
}
\description{
Check Built-in Variable Importance Availability for caret Models
}
