% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn15_all_models.R
\name{combine_model_roc_results}
\alias{combine_model_roc_results}
\title{Combine Model ROC Results}
\usage{
combine_model_roc_results(object)
}
\arguments{
\item{object}{PrognosiX object.}
}
\value{
Data frame of combined results.
}
\description{
Combines ROC results from all trained models in a PrognosiX object.
}
