% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn12_coxph_model.R
\name{compute_hr_coxph}
\alias{compute_hr_coxph}
\title{Compute HR for CoxPH}
\usage{
compute_hr_coxph(fit_coxph)
}
\arguments{
\item{fit_coxph}{CoxPH model.}
}
\value{
HR results.
}
\description{
Computes HR and CI for CoxPH coefficients.
}
