% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn8_lasso_model.R
\name{compute_hr_lasso}
\alias{compute_hr_lasso}
\title{Compute HR for LASSO}
\usage{
compute_hr_lasso(
  lasso_model,
  time_col = "time",
  status_col = "status",
  lambda = NULL
)
}
\arguments{
\item{lasso_model}{LASSO model.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{lambda}{Lambda.}
}
\value{
HR results data frame.
}
\description{
Computes HR and CI for LASSO coefficients.
}
