% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn10_pls_model.R
\name{compute_hr_pls}
\alias{compute_hr_pls}
\title{Compute HR for PLS}
\usage{
compute_hr_pls(pls_model, train_data)
}
\arguments{
\item{pls_model}{PLS model.}

\item{train_data}{Training data.}
}
\value{
HR results data frame.
}
\description{
Computes HR and CI for PLS coefficients.
}
