% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_create_survival_time.R
\name{convert_two_class_to_binary}
\alias{convert_two_class_to_binary}
\title{Convert Two Class to Binary}
\usage{
convert_two_class_to_binary(data)
}
\arguments{
\item{data}{Data frame.}
}
\description{
Convert Two Class to Binary
}
