% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_04_descriptive.R, R/utils_convert_types.R
\name{convert_variables}
\alias{convert_variables}
\title{Convert Variables to Numeric or Factor Based on Their Types}
\usage{
convert_variables(
  data,
  variable_types,
  save_dir = NULL,
  save_data = F,
  group_col = NULL,
  csv_filename = "clean_data.csv"
)

convert_variables(
  data,
  variable_types,
  save_dir = NULL,
  save_data = F,
  group_col = NULL,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{data}{A data frame or an S4 object (Stat, Train_Model, Subtyping, PrognosiX).}

\item{variable_types}{A list containing variable type information.}

\item{save_dir}{Directory to save the cleaned data.}

\item{save_data}{Logical. Whether to save the data.}

\item{group_col}{The name of the group column.}

\item{csv_filename}{The name of the CSV file.}
}
\value{
A data frame with the variables converted to the appropriate types
  (numeric or factor).
}
\description{
This function converts the variables (columns) in the input data frame to either
numeric or factor types based on the information provided in `variable_types`.
Numeric variables are converted to numeric data type, and non-numeric variables
are converted to factors. The function uses `variable_types` to determine which
variables should be treated as numeric and which should be treated as factors.
}
\examples{
data(iris)
types <- diagnose_variable_type(iris, group_col = "Species")
converted <- convert_variables(iris, types, save_data = FALSE)
str(converted)
\dontrun{
types <- diagnose_variable_type(stat_obj_test@clean.data, group_col = "SWAB")
converted <- convert_variables(stat_obj_test@clean.data, types, 
group_col = "SWAB", save_data = FALSE)
}
}
