% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-3.R
\name{cor_top_correlations}
\alias{cor_top_correlations}
\title{Correlation Analysis and Plot for Top Correlated Features}
\usage{
cor_top_correlations(
  object,
  correlation_threshold = 0.9,
  top_n = 15,
  palette_name = "Zissou1",
  data_type = "scale",
  save_plots = TRUE,
  plot_width = 5,
  plot_height = 5,
  save_dir = here("StatObject", "correlation_info"),
  base_size = 14
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame containing the features to be analyzed.}

\item{correlation_threshold}{A numeric value specifying the correlation threshold for feature selection (default is 0.9).}

\item{top_n}{An integer specifying the number of top correlated features to be plotted (default is 15).}

\item{palette_name}{A string specifying the color palette for the plot.}

\item{data_type}{A string specifying the type of data to extract from the 'Stat' object. Either "clean" or "scale" (default is "scale").}

\item{save_plots}{A logical value indicating whether to save the plot as a PDF (default is TRUE).}

\item{plot_width}{A numeric value specifying the width of the saved plot (default is 5).}

\item{plot_height}{A numeric value specifying the height of the saved plot (default is 5).}

\item{save_dir}{A string specifying the directory where the plot will be saved (default is \code{here("StatObject", "correlation_info")}).}

\item{base_size}{A numeric value specifying the base font size for the plot (default is 14).}
}
\value{
If the input is a 'Stat' object, the updated object with the correlation plot added to the \code{corr.result} slot.
Otherwise, the function returns the plot for the top correlated features.
}
\description{
This function extracts data from a given object or data frame, performs a correlation analysis
on the features, and then generates a plot for the top \code{n} most correlated features. The user can
specify a correlation threshold, the number of top correlated features to plot, and customize the plot settings.
If the input object is of class "Stat", the function updates the \code{corr.result} slot of the object with the plot.
}
\examples{
# Perform correlation analysis and plot top 10 correlated features
cor_top_correlations(my_stat_object, top_n = 10)

# Perform analysis and save the plot of top 20 correlated features with custom plot dimensions
cor_top_correlations(my_stat_object, top_n = 20, plot_width = 7, plot_height = 6)

}
