% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn11_coxboost_model.R
\name{coxboost_coefs_compute_hr_and_ci}
\alias{coxboost_coefs_compute_hr_and_ci}
\title{Compute HR and CI for CoxBoost Coefficients}
\usage{
coxboost_coefs_compute_hr_and_ci(object)
}
\arguments{
\item{object}{PrognosiX object or model.}
}
\value{
Updated PrognosiX object or results.
}
\description{
Computes HR and CI for CoxBoost coefficients.
}
