% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn12_coxph_model.R
\name{coxph_compute_hr_and_ci}
\alias{coxph_compute_hr_and_ci}
\title{Compute HR and CI for CoxPH (Wrapper)}
\usage{
coxph_compute_hr_and_ci(object)
}
\arguments{
\item{object}{PrognosiX object.}
}
\value{
Updated PrognosiX object or results.
}
\description{
Wrapper to compute HR/CI for CoxPH in PrognosiX.
}
