% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_create_forest_plot.R,
%   R/module_4_pn5_1_forest_plot.R
\name{create_forest_plot}
\alias{create_forest_plot}
\title{Create Forest Plot}
\usage{
create_forest_plot(
  hr_results,
  plot_title =
    "Univariate Analysis: Clinical Factors and Hazard Ratios with 95\% Confidence Intervals",
  save_plot = FALSE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 11,
  plot_height = 5,
  palette_name = "AsteroidCity1",
  base_size = 14,
  hr_limit = c(0, 3),
  ci_range_limit = 1000
)

create_forest_plot(
  hr_results,
  plot_title =
    "Univariate Analysis: Clinical Factors and Hazard Ratios with 95\% Confidence Intervals",
  save_plot = FALSE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 11,
  plot_height = 5,
  palette_name = "AsteroidCity1",
  base_size = 14,
  hr_limit = c(0, 3),
  ci_range_limit = 1000
)
}
\arguments{
\item{hr_results}{Data frame containing HR results (Variable, HR, CI_lower, CI_upper, P_value, HR_95CI).}

\item{plot_title}{Title of the plot.}

\item{save_plot}{Logical, whether to save the plot.}

\item{save_dir}{Directory to save the plot.}

\item{plot_width}{Width of the saved plot.}

\item{plot_height}{Height of the saved plot.}

\item{palette_name}{Color palette name (from wesanderson package).}

\item{base_size}{Base font size.}

\item{hr_limit}{Numeric vector of length 2 defining the x-axis limits for HR.}

\item{ci_range_limit}{Maximum allowed range for CI (to filter out extreme intervals).}
}
\value{
A combined ggplot object.
}
\description{
Creates a forest plot visualizing Hazard Ratios and Confidence Intervals.
}
