% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn8_lasso_model.R
\name{create_forest_plot_lasso}
\alias{create_forest_plot_lasso}
\title{Create Forest Plot for LASSO Model}
\usage{
create_forest_plot_lasso(
  hr_results,
  plot_title = "Evaluation of Hazard Ratios and Confidence Intervals",
  save_plot = FALSE,
  save_dir = here::here("Prognosi"),
  plot_width = 11,
  plot_height = 5,
  palette_name = "AsteroidCity1",
  base_size = 14,
  hr_limit = c(0, 3),
  ci_range_limit = 1000
)
}
\arguments{
\item{hr_results}{HR results.}

\item{plot_title}{Title.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{palette_name}{Palette.}

\item{base_size}{Size.}

\item{hr_limit}{HR limit.}

\item{ci_range_limit}{CI range limit.}
}
\value{
Plot object.
}
\description{
Creates a forest plot for LASSO model HR results.
}
