% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn14_rsf_model.R
\name{create_forest_plot_rsf}
\alias{create_forest_plot_rsf}
\title{Create Forest Plot for RSF}
\usage{
create_forest_plot_rsf(
  hr_results,
  plot_title = "Variable Importance for RSF Model",
  save_plot = FALSE,
  save_dir = here::here("Prognosi"),
  plot_width = 11,
  plot_height = 5,
  palette_name = "AsteroidCity1",
  base_size = 14,
  hr_limit = c(0, 3),
  ci_range_limit = 1000
)
}
\arguments{
\item{hr_results}{HR results.}

\item{plot_title}{Title.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{palette_name}{Palette.}

\item{base_size}{Size.}

\item{hr_limit}{HR limit.}

\item{ci_range_limit}{CI limit.}
}
\value{
Plot.
}
\description{
Internal function to create forest plot for RSF (Importance plot).
}
