% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_01_object.R
\name{CreateStatObject}
\alias{CreateStatObject}
\title{Create a Stat object for statistical analysis}
\usage{
CreateStatObject(
  raw.data = data.frame(),
  clean.data = data.frame(),
  info.data = data.frame(),
  group_col = "group",
  na.action = c("allow", "omit", "error"),
  ...
)
}
\arguments{
\item{raw.data}{A data.frame containing the raw data for analysis. Defaults to an empty data frame.}

\item{clean.data}{A data.frame containing the cleaned data for analysis. Defaults to an empty data frame.}

\item{info.data}{A data.frame containing additional metadata related to the data. Defaults to an empty data frame.}

\item{group_col}{A character string specifying the name of the column used for grouping. Default is `"group"`.}

\item{na.action}{Character string specifying how to handle NA values. 
Options are "allow" (keep NA values, default), "omit" (remove rows with NA),
or "error" (stop if NA values are found).}

\item{...}{Additional arguments passed to methods (not used in this function).}
}
\value{
An object of class `Stat`, which contains the processed data and metadata as slots.
}
\description{
This function creates a `Stat` object, which is used to store and manage various components
of statistical analysis, such as raw data, cleaned data, additional metadata, and processing
information. It performs basic checks and preparation on the input data before creating the object.
}
\examples{
raw <- data.frame(id = 1:10, value = rnorm(10), group = rep(c("A","B"), each=5))
inf<-data.frame(age = 1:10, value = rnorm(10), group = rep(c("A","B"), each=5))
stat_new <- CreateStatObject(raw.data = raw,  group_col = "group")
print(stat_new)
stat_new <- CreateStatObject(raw.data = raw, clean.data = raw, group_col = "group")
print(stat_new)
stat_new <- CreateStatObject(raw.data = raw, group_col = "group",info.data = inf)
print(stat_new)
}
