% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_create_surv_obj.R
\name{CreateSurObj}
\alias{CreateSurObj}
\title{Create SurObj}
\usage{
CreateSurObj(
  clean.data = NULL,
  info.data = data.frame(),
  time_col = "time",
  status_col = "status",
  clustered.data = data.frame(),
  survival.data = data.frame(),
  sub.data = data.frame(),
  baseline.table = NULL,
  univariate.analysis = list(),
  survival.analysis = list(),
  rcss.results = NULL,
  data_type = "train",
  object = NULL
)
}
\arguments{
\item{clean.data}{Clean data.}

\item{info.data}{Info data.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{clustered.data}{Clustered data.}

\item{survival.data}{Survival data.}

\item{sub.data}{Sub data.}

\item{baseline.table}{Baseline table.}

\item{univariate.analysis}{Univariate analysis.}

\item{survival.analysis}{Survival analysis.}

\item{rcss.results}{RCSS results.}

\item{data_type}{Data type.}

\item{object}{Input object.}
}
\description{
Create SurObj
}
