% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-3.R
\name{cross_analysis}
\alias{cross_analysis}
\title{Cross-Analysis of Two Variables with Scatter Plot and Correlation Test}
\usage{
cross_analysis(
  data,
  vars,
  group_col = "group",
  palette_name = "Zissou1",
  save_plots = TRUE,
  plot_width = 5,
  plot_height = 5,
  save_dir = here("StatObject", "correlation_info")
)
}
\arguments{
\item{data}{A data frame containing the variables to be analyzed.}

\item{vars}{A character vector of length 2 specifying the names of the two variables for the analysis.}

\item{group_col}{A character string specifying the name of the grouping column for coloring points in the plot (default is "group").}

\item{palette_name}{A string specifying the color palette to be used for the plot.}

\item{save_plots}{A logical value indicating whether to save the plot as a PDF file (default is TRUE).}

\item{plot_width}{A numeric value specifying the width of the saved plot in inches (default is 5).}

\item{plot_height}{A numeric value specifying the height of the saved plot in inches (default is 5).}

\item{save_dir}{A string specifying the directory where the plot will be saved (default is \code{here("StatObject", "correlation_info")}).}
}
\value{
A scatter plot of the two specified variables, with the correlation coefficient and p-value annotated.
If \code{save_plots} is TRUE, the plot is saved as a PDF file.
}
\description{
This function performs a cross-analysis of two numeric variables in a given data frame. It generates
a scatter plot of the two variables, fits a linear regression line, and displays the correlation coefficient
along with the p-value. The user can customize the plot appearance and save the plot as an image file.
}
\examples{
# Perform cross-analysis of two variables 'var1' and 'var2'
cross_analysis(data = my_data, vars = c("var1", "var2"))

# Perform cross-analysis with a custom plot size and save the plot
cross_analysis(data = my_data, vars = c("var1", "var2"), plot_width = 6, plot_height = 6)

}
