% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-3.R
\name{cross_plot_analysis}
\alias{cross_plot_analysis}
\title{Cross-Plot Analysis of Two Variables from a 'Stat' Object or Data Frame}
\usage{
cross_plot_analysis(
  object,
  vars,
  palette_name = "Zissou1",
  data_type = "scale",
  group_col = "group",
  save_plots = TRUE,
  plot_width = 5,
  plot_height = 5,
  save_dir = here("StatObject", "correlation_info")
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame containing the data for analysis.}

\item{vars}{A character vector of length 2 specifying the names of the two variables for the analysis.}

\item{palette_name}{A string specifying the color palette for the plot.}

\item{data_type}{A string indicating which type of data to extract from the 'Stat' object. Options are "clean" or "scale" (default is "scale").}

\item{group_col}{A string specifying the column name for the grouping variable (default is "group").}

\item{save_plots}{A logical value indicating whether to save the plot as a PDF file (default is TRUE).}

\item{plot_width}{A numeric value specifying the width of the saved plot in inches (default is 5).}

\item{plot_height}{A numeric value specifying the height of the saved plot in inches (default is 5).}

\item{save_dir}{A string specifying the directory where the plot will be saved (default is \code{here("StatObject", "correlation_info")}).}
}
\value{
A scatter plot of the two specified variables, with the correlation coefficient and p-value annotated.
If \code{save_plots} is TRUE, the plot is saved as a PDF file.
}
\description{
This function performs a cross-plot analysis of two variables by generating a scatter plot with a correlation test.
It handles input either as a 'Stat' object or a data frame, extracting the relevant data based on the specified \code{data_type}.
The plot will be saved as a PDF file if specified.
}
\examples{
# Perform cross-plot analysis on a 'Stat' object with specified variables
cross_plot_analysis(object = my_stat_object, vars = c("var1", "var2"))

# Perform cross-plot analysis on a data frame with a custom plot size
cross_plot_analysis(object = my_data, vars = c("var1", "var2"), plot_width = 6, plot_height = 6)

}
