% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_07_evaluate.R
\name{davies_bouldin}
\alias{davies_bouldin}
\title{Davies-Bouldin Index}
\usage{
davies_bouldin(clustered.data, label_col = "group")
}
\arguments{
\item{clustered.data}{Clustered data.}

\item{label_col}{Column for label.}
}
\value{
Numeric DB index value.
}
\description{
Davies-Bouldin Index
}
\examples{
\dontrun{
  df <- data.frame(x = rnorm(30), y = rnorm(30), group = sample(1:3, 30, replace = TRUE))
  db <- davies_bouldin(df)
  cat("Davies-Bouldin Index:", db, "\n")
}
}
