% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_10_deploy.R
\name{deploy_clinlab_app}
\alias{deploy_clinlab_app}
\title{Clinlabomics Terminal: Light-Tech Edition}
\usage{
deploy_clinlab_app(deployment, title = "Clinlabomics Terminal")
}
\arguments{
\item{deployment}{A ModelDeployment object.}

\item{title}{Character string for the application window title.}
}
\description{
Futuristic UI with high-contrast elements on a light grey base for clinical model deployment.
}
