% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn17_extra_val.R
\name{evaluate_best_model_val}
\alias{evaluate_best_model_val}
\title{Evaluate Best Model Validation}
\usage{
evaluate_best_model_val(
  object,
  validation_data,
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "best_model_val"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{validation_data}{Validation data.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Size.}
}
\value{
Updated PrognosiX object.
}
\description{
Evaluates the best model on validation data.
}
