% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn12_coxph_model.R
\name{evaluate_coxph_km}
\alias{evaluate_coxph_km}
\title{Evaluate CoxPH Kaplan-Meier}
\usage{
evaluate_coxph_km(
  fit_best_cv_coxph,
  data,
  data_name = "test",
  time_col = "time",
  status_col = "status",
  R = 1000,
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "coxph_model"),
  palette_name = "Dark2",
  plot_width = 10,
  plot_height = 8,
  base_size = 14,
  seed = 1234,
  fill_color = "lightblue"
)
}
\arguments{
\item{fit_best_cv_coxph}{Model.}

\item{data}{Data.}

\item{data_name}{Data name.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{R}{Resampling.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{palette_name}{Palette.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Size.}

\item{seed}{Seed.}

\item{fill_color}{Fill color.}
}
\value{
List with KM results and plots.
}
\description{
Evaluates KM curves for CoxPH risk groups.
}
