% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn12_coxph_model.R
\name{evaluate_coxph_roc}
\alias{evaluate_coxph_roc}
\title{Evaluate CoxPH ROC}
\usage{
evaluate_coxph_roc(
  fit_coxph_model,
  train_data = NULL,
  test_data = NULL,
  validation_data = NULL,
  palette_name = "AsteroidCity1",
  time_col = "time",
  status_col = "status",
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "coxph_model"),
  plot_width = 7,
  plot_height = 7,
  base_size = 14
)
}
\arguments{
\item{fit_coxph_model}{CoxPH model.}

\item{train_data}{Training data.}

\item{test_data}{Testing data.}

\item{validation_data}{Validation data.}

\item{palette_name}{Palette.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Size.}
}
\value{
List with ROC results and plot.
}
\description{
Evaluates CoxPH model performance using ROC curves.
}
