% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn8_lasso_model.R
\name{evaluate_km_lasso_model}
\alias{evaluate_km_lasso_model}
\title{Evaluate KM for LASSO Model (Wrapper)}
\usage{
evaluate_km_lasso_model(
  object,
  time_col = "time",
  status_col = "status",
  nfolds = 10,
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "lasso_model"),
  plot_width = 7,
  plot_height = 7,
  base_size = 14,
  data_name = "test",
  data = NULL
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{nfolds}{Folds.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Size.}

\item{data_name}{Data name.}

\item{data}{Data.}
}
\value{
Updated PrognosiX object or results.
}
\description{
Wrapper to evaluate KM for LASSO model in PrognosiX object.
}
