% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn8_lasso_model.R
\name{evaluate_lasso_km}
\alias{evaluate_lasso_km}
\title{Evaluate LASSO Kaplan-Meier}
\usage{
evaluate_lasso_km(
  fit_best_cv_lasso,
  best_lambda_value,
  data,
  data_name = "test",
  time_col = "time",
  status_col = "status",
  R = 1000,
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "lasso_model"),
  palette_name = "Dark2",
  plot_width = 10,
  plot_height = 8,
  base_size = 14,
  seed = 1234,
  fill_color = "lightblue"
)
}
\arguments{
\item{fit_best_cv_lasso}{Model.}

\item{best_lambda_value}{Lambda.}

\item{data}{Data.}

\item{data_name}{Data name ("test", "val").}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{R}{Resampling (unused).}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{palette_name}{Palette.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Size.}

\item{seed}{Seed.}

\item{fill_color}{Fill color.}
}
\value{
List with KM results and plots.
}
\description{
Evaluates KM curves for LASSO risk groups.
}
