% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn11_coxboost_model.R
\name{evaluate_roc_coxboost_model}
\alias{evaluate_roc_coxboost_model}
\title{Evaluate ROC for CoxBoost Model (Wrapper)}
\usage{
evaluate_roc_coxboost_model(
  object,
  time_col = "time",
  status_col = "status",
  palette_name = "AsteroidCity1",
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "coxboost_model"),
  plot_width = 7,
  plot_height = 7,
  base_size = 14
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{palette_name}{Palette.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Size.}
}
\value{
Updated PrognosiX object or results.
}
\description{
Wrapper to evaluate ROC for CoxBoost model in PrognosiX.
}
