% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-9.R
\name{Extract_external_validata}
\alias{Extract_external_validata}
\title{Extract External Validation Data and Update Model}
\usage{
Extract_external_validata(
  data = NULL,
  object_stats = NULL,
  object_model = NULL,
  group_col = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing the external validation data. Must not have missing values.}

\item{object_stats}{An object of class 'Stat' from which clean validation data will be extracted. Must not have missing values.}

\item{object_model}{A 'Model_data' object that will be updated with the external validation data.}

\item{group_col}{The column name in the validation data that indicates the group or outcome. (Optional)}

\item{...}{Additional parameters that may be passed to other functions (if any).}
}
\value{
The updated 'Model_data' object, with the 'filtered.set' slot containing the external validation data.
}
\description{
This function extracts an external validation dataset either from a provided data frame or from an existing 'Stat' object,
processes it (e.g., handles missing values and converts factors), and updates the provided model object with the validation data.
}
\examples{
updated_model <- Extract_external_validata(data = validation_data,
                                          object_model = model_object)
}
