% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn15_all_models.R
\name{extract_model_results}
\alias{extract_model_results}
\title{Extract Model Results}
\usage{
extract_model_results(object, metric = "C_index")
}
\arguments{
\item{object}{PrognosiX object.}

\item{metric}{Metric to optimize ("C_index" or "ROC_AUC").}
}
\value{
Updated PrognosiX object.
}
\description{
Extracts and identifies the best model based on a metric.
}
