% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_06_deg.R
\name{ExtractLastTestSig}
\alias{ExtractLastTestSig}
\title{Extract Last Significant Test Results}
\usage{
ExtractLastTestSig(object)
}
\arguments{
\item{object}{An object of class `Stat`. The function attempts to extract the `last_test_sig` from the `var.result` slot.}
}
\value{
Returns the last significant test results stored in the `last_test_sig` slot of the `var.result` list.
If the `last_test_sig` is not found, it returns `NULL`.
}
\description{
This function extracts the last significant test results stored in the `var.result` slot of a `Stat` object. If
the `last_test_sig` result is not available, it returns `NULL`.
}
\examples{
\dontrun{
data(iris)
iris_sub <- iris[iris$Species != "virginica", ]
stat <- CreateStatObject(clean.data = iris_sub, group_col = "Species")
stat <- stat_var_feature(stat, p_threshold = 0.05)
last <- ExtractLastTestSig(stat)
head(last)
}
}
