% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn17_extra_val.R
\name{ExtractPrognosiXValidata}
\alias{ExtractPrognosiXValidata}
\title{Extract Validation Data}
\usage{
ExtractPrognosiXValidata(
  data = NULL,
  object_stats = NULL,
  object_prognosix = NULL,
  group_col = NULL,
  time_col = "time",
  status_col = "status",
  ...
)
}
\arguments{
\item{data}{Data frame.}

\item{object_stats}{Stat object.}

\item{object_prognosix}{PrognosiX object.}

\item{group_col}{Group column.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{...}{Extra args.}
}
\value{
Updated PrognosiX object.
}
\description{
Extracts validation data for PrognosiX object.
}
