% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_08_normalize.R
\name{ExtractScaleData}
\alias{ExtractScaleData}
\title{Extract Scaled Data}
\usage{
ExtractScaleData(object)
}
\arguments{
\item{object}{Stat object.}
}
\description{
Extract Scaled Data
}
\examples{
stat <- CreateStatObject(clean.data = mtcars, group_col = "cyl")
stat <- stat_normalize_process(stat, method = "z_score")
scaled <- ExtractScaleData(stat)
head(scaled)
}
