% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_cutsubgroup.R
\name{find_maxstat_cutoff}
\alias{find_maxstat_cutoff}
\title{Find MaxStat Cutoff}
\usage{
find_maxstat_cutoff(
  data,
  variables,
  time_col = "time",
  status_col = "status",
  n_iterations = 10000,
  min_group_size = 0.3,
  seed = 123
)
}
\arguments{
\item{data}{Data frame.}

\item{variables}{Variables.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{n_iterations}{Iterations.}

\item{min_group_size}{Min group size.}

\item{seed}{Seed.}
}
\description{
Find MaxStat Cutoff
}
