% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_cutsubgroup.R
\name{find_roc_cutoff}
\alias{find_roc_cutoff}
\title{Find ROC Cutoff}
\usage{
find_roc_cutoff(
  data,
  variables,
  time_col = "time",
  status_col = "status",
  seed = 123
)
}
\arguments{
\item{data}{Data frame.}

\item{variables}{Variables.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{seed}{Seed.}
}
\description{
Find ROC Cutoff
}
