% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn10_pls_model.R
\name{forest_plot_pls_model}
\alias{forest_plot_pls_model}
\title{Forest Plot for PLS Model (Wrapper)}
\usage{
forest_plot_pls_model(
  object,
  plot_title =
    "Evaluation of Hazard Ratios and Confidence Intervals for pls Model in Survival Analysis",
  time_col = "time",
  status_col = "status",
  var_col = NULL,
  palette_name = "AsteroidCity1",
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "pls_model"),
  plot_width = 14,
  plot_height = 7,
  base_size = 14,
  use_subgroup_data = FALSE,
  hr_limit = c(0.1, 3)
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{plot_title}{Title.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{var_col}{Variable column.}

\item{palette_name}{Palette.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Size.}

\item{use_subgroup_data}{Logical.}

\item{hr_limit}{HR limit.}
}
\value{
Updated PrognosiX object or plot.
}
\description{
Generates forest plot for PLS model in PrognosiX.
}
