% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub4_nmf.R
\name{generate_consensus_map_plot}
\alias{generate_consensus_map_plot}
\title{Generate Consensus Map Plot}
\usage{
generate_consensus_map_plot(
  estimate,
  save_dir = here("Subtyping", "cluster_results", "nmf_results")
)
}
\arguments{
\item{estimate}{NMF estimate.}

\item{save_dir}{Directory.}
}
\description{
Generate Consensus Map Plot
}
