% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub4_nmf.R
\name{generate_nmf_plot}
\alias{generate_nmf_plot}
\title{Generate NMF Plot}
\usage{
generate_nmf_plot(
  estimate,
  save_dir = here("Subtyping", "cluster_results", "nmf_results")
)
}
\arguments{
\item{estimate}{NMF estimate.}

\item{save_dir}{Directory.}
}
\description{
Generate NMF Plot
}
