% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config.R
\name{get_output_root}
\alias{get_output_root}
\title{Get Global Output Root Directory}
\usage{
get_output_root()
}
\value{
Character string with the current output root (default: ./icare_output)
}
\description{
Get Global Output Root Directory
}
\examples{
\dontrun{
get_output_root()
}
}
