% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-3.R
\name{highly_correlated_features}
\alias{highly_correlated_features}
\title{Identify Highly Correlated Features and Generate Heatmap}
\usage{
highly_correlated_features(
  data,
  correlation_threshold = 0.95,
  palette_name = "Zissou1",
  save_plots = TRUE,
  save_dir = here("StatObject", "correlation_info"),
  heatmap_title = "Correlation Heatmap",
  plot_width = 10,
  plot_height = 8,
  save_data = TRUE,
  csv_filename = "correlation_heatmap_data.csv"
)
}
\arguments{
\item{data}{A data frame containing the features to be analyzed.}

\item{correlation_threshold}{A numeric value specifying the threshold for correlation. Features with a correlation
greater than this value will be considered correlated (default is 0.95).}

\item{palette_name}{A string specifying the color palette to be used for the heatmap}

\item{save_plots}{A logical value indicating whether to save the heatmap plot (default is TRUE).}

\item{save_dir}{The directory where the heatmap plot will be saved, if \code{save_plots} is TRUE (default is \code{here("StatObject", "correlation_info")}).}

\item{heatmap_title}{A string specifying the title of the heatmap (default is "Correlation Heatmap").}
}
\value{
A list containing the following components:
\itemize{
\item \code{corr_matrix}: The correlation matrix of the features.
\item \code{correlated_features}: A character vector of the features that are highly correlated.
\item \code{record_collinear}: A data frame with pairs of correlated features and their correlation values.
\item \code{top5_pairs}: A data frame containing the top 5 most correlated feature pairs.
\item \code{original_data}: The original data frame passed to the function.
}
}
\description{
This function identifies highly correlated features in a data frame and generates a heatmap to visualize
the correlation matrix. It also returns a list of the correlated features and the top 5 most correlated feature pairs.
The function can optionally save the heatmap plot to a specified directory.
}
\examples{
# Assuming 'data' is a data frame of features
result <- highly_correlated_features(data, correlation_threshold = 0.95, save_plots = TRUE)
head(result$top5_pairs)

}
