% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_05_handle_missing.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute Missing Values in Data}
\usage{
impute_missing_values(
  data,
  group_col = "group",
  impute_method = "mice",
  m = 5,
  max_unique_values = 5,
  return_imputation_info = T,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{data}{A data frame.}

\item{group_col}{Group column.}

\item{impute_method}{Method ("mice" or "median_mode").}

\item{m}{Number of imputations.}

\item{max_unique_values}{Max unique values for categorical.}

\item{return_imputation_info}{Logical.}

\item{save_dir}{Save directory.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Impute Missing Values in Data
}
\examples{
data(iris)
iris_na <- iris
iris_na[1:5, 1] <- NA
imp <- impute_missing_values(iris_na, impute_method = "median_mode", return_imputation_info = FALSE)
head(imp)
}
