% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn8_lasso_model.R
\name{lasso_compute_hr_and_ci}
\alias{lasso_compute_hr_and_ci}
\title{Compute HR and CI for LASSO (Wrapper)}
\usage{
lasso_compute_hr_and_ci(
  object,
  time_col = "time",
  status_col = "status",
  lambda = NULL
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{lambda}{Lambda.}
}
\value{
Updated PrognosiX or results.
}
\description{
Wrapper to compute HR/CI for LASSO in PrognosiX.
}
