% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn6_feature_selection.R
\name{lasso_cv_visualization}
\alias{lasso_cv_visualization}
\title{Visualize LASSO Cross-Validation}
\usage{
lasso_cv_visualization(
  object,
  palette_name = "AsteroidCity1",
  base_size = 14,
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "Sel_feature"),
  plot_width = 5,
  plot_height = 5
)
}
\arguments{
\item{object}{PrognosiX object or model list.}

\item{palette_name}{Palette name.}

\item{base_size}{Base font size.}

\item{save_plots}{Logical.}

\item{save_dir}{Save directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}
}
\value{
Plot object.
}
\description{
plots the cross-validation curve for LASSO.
}
