% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn6_feature_selection.R
\name{lasso_feature_selection}
\alias{lasso_feature_selection}
\title{LASSO Feature Selection}
\usage{
lasso_feature_selection(data, time_col, status_col, alpha = 1, seed = 123)
}
\arguments{
\item{data}{Data frame.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{alpha}{Alpha parameter for glmnet (1 for LASSO).}

\item{seed}{Seed.}
}
\value{
List containing model, best lambda, important variables, etc.
}
\description{
Performs feature selection using LASSO Cox regression.
}
