% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_03_deploy.R
\name{launch_prog_deploy_app}
\alias{launch_prog_deploy_app}
\title{Launch the interactive Prognosis Terminal}
\usage{
launch_prog_deploy_app(
  prog_manager,
  title = NULL,
  var_dict = NULL,
  project_info = NULL
)
}
\arguments{
\item{prog_manager}{A \code{Prog_Manager} object}

\item{title}{Browser title (overrides custom text)}

\item{var_dict}{Data frame with columns Feature, Description, Units}

\item{project_info}{List with elements \code{abstract} and \code{citation}}
}
\value{
Runs the Shiny app
}
\description{
Launch the interactive Prognosis Terminal
}
\examples{
\dontrun{
# Requires a Prog_Manager object
# mgr <- New_Prog_Manager(prog_obj)
# launch_prog_deploy_app(mgr)
}
}
