% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{list_presets}
\alias{list_presets}
\title{List all available presets (both old and new)}
\usage{
list_presets()
}
\value{
Prints a table with preset names, type, and description.
}
\description{
List all available presets (both old and new)
}
