% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{match_factor_levels}
\alias{match_factor_levels}
\title{Match factor levels of two data frames}
\usage{
match_factor_levels(data, ref)
}
\description{
Match factor levels of two data frames
}
\keyword{internal}
