% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-10-sel-feature.R
\name{mic_information}
\alias{mic_information}
\title{Calculate Maximal Information Coefficients (MIC) Between Features and Target}
\usage{
mic_information(x, y)
}
\arguments{
\item{x}{A data frame or matrix containing the features (predictor variables).}

\item{y}{A numeric vector representing the target variable (response).}
}
\value{
A data frame with two columns:
\describe{
\item{\code{Feature}}{The names of the features (columns of \code{x}).}
\item{\code{Maximal Information Coefficient}}{The corresponding MIC values between each feature and \code{y}.}
}
}
\description{
This function computes the Maximal Information Coefficient (MIC) between each column in the feature matrix \code{x}
and the target vector \code{y}, using the \code{mine} function from the \code{minerva} package.
}
\examples{
\dontrun{
x <- data.frame(
  var1 = rnorm(100),
  var2 = runif(100),
  var3 = rpois(100, lambda = 3)
)
y <- rnorm(100)
mic_info <- mic_information(x, y)
print(mic_info)
}
}
