% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-8.R
\name{ModelBestRoc}
\alias{ModelBestRoc}
\title{Model Evaluation and ROC Curve Plotting for Best Model}
\usage{
ModelBestRoc(
  object,
  group_col = "group",
  palette_name = "AsteroidCity1",
  base_size = 14,
  save_plots = TRUE,
  save_dir = here("ModelData", "best_model_result"),
  plot_width = 5,
  plot_height = 5,
  alpha = 1,
  metric = "auc",
  metric_function = max
)
}
\arguments{
\item{object}{An object of class 'Model_data' containing training models and evaluation results.}

\item{group_col}{The name of the label column in the dataset, default is "group".}

\item{palette_name}{The name of the color palette to use in the ROC plot, default is "AsteroidCity1".}

\item{base_size}{The base font size for the plot, default is 14.}

\item{save_plots}{Logical value indicating whether to save the plot. Default is TRUE.}

\item{save_dir}{The directory where the plot will be saved, default is "ModelData/best_model_result".}

\item{plot_width}{The width of the saved plot, default is 5.}

\item{plot_height}{The height of the saved plot, default is 5.}

\item{alpha}{The transparency of plot elements, default is 1 (completely opaque).}

\item{metric}{The metric used to select the best model (e.g., "auc", "accuracy"). Default is "auc".}

\item{metric_function}{The function used to determine the best model based on the metric (e.g., \code{max} for AUC, \code{min} for error). Default is \code{max}.}
}
\value{
The updated 'Model_data' object, with the 'best.model.result' slot updated with the best model and its test result.
}
\description{
This function identifies the best model based on a specified performance metric from an object of class 'Model_data',
evaluates its performance on the test dataset, and plots the ROC curve for both training and testing datasets.
Optionally, the plot can be saved to a specified directory.
}
\examples{
updated_object <- ModelBestRoc(object = model_data_object,
                               group_col = "group",
                               metric = "auc",
                               metric_function = max)
}
