% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-14.R
\name{ModelThreshold}
\alias{ModelThreshold}
\title{Model Threshold Selection and Update for Model_data Object}
\usage{
ModelThreshold(
  object,
  group_col = "group",
  threshold_metrics = NULL,
  method_threshold = c("max_accuracy", "closest_ppv", "closest_npv"),
  title = NULL,
  theme = theme_scp(),
  palette_name = "Zissou1",
  base_size = 14,
  save_plots = TRUE,
  save_dir = here("ModelData", "best_model_result"),
  plot_width = 5,
  plot_height = 5
)
}
\arguments{
\item{object}{An object of class 'Model_data', which contains the model and other necessary data.}

\item{group_col}{A character string specifying the name of the column representing the group
(or class) in the testing dataset. Default is "group".}

\item{threshold_metrics}{A data frame containing threshold metrics (e.g., Accuracy, PPV, NPV)
calculated for different thresholds. If \code{NULL}, the metrics will be calculated within the function.}

\item{method_threshold}{A character string specifying the method for selecting the best threshold.
It can be one of: "max_accuracy", "closest_ppv", or "closest_npv". Default is "max_accuracy".}

\item{title}{A character string specifying the title of the plot. If \code{NULL}, a default title
will be used based on the method selected. Default is \code{NULL}.}

\item{theme}{A ggplot2 theme object for customizing the plot appearance. Default is \code{theme_scp()}.}

\item{palette_name}{A character string specifying the color palette for the plot. Default is "Zissou1".}

\item{base_size}{An integer specifying the base font size for the plot elements. Default is 14.}

\item{save_plots}{A logical value indicating whether to save the plot. Default is TRUE.}

\item{save_dir}{A character string specifying the directory where the plot will be saved.
Default is \code{"ModelData/best_model_result"}.}

\item{plot_width}{An integer specifying the width of the saved plot in inches. Default is 5.}

\item{plot_height}{An integer specifying the height of the saved plot in inches. Default is 5.}
}
\value{
An updated \code{Model_data} object with the best threshold and associated information
stored in the \code{best.model.result} slot.
}
\description{
This function selects the best threshold for model performance based on the specified
method, and updates the input \code{Model_data} object with the selected threshold and
associated metrics. The function predicts probabilities using the model in the
\code{Model_data} object, calculates the corresponding threshold metrics, and selects
the best threshold using one of the provided methods ("max_accuracy", "closest_ppv",
or "closest_npv"). The function also generates and saves relevant plots if desired.
}
\examples{
\dontrun{
# Example usage:
model_data_object <- ModelThreshold(object = model_data_object,
                                    method_threshold = "max_accuracy")
}
}
