% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_03_deploy.R
\name{New_Prog_Manager}
\alias{New_Prog_Manager}
\title{Create a deployment manager}
\usage{
New_Prog_Manager(prog_train_obj)
}
\arguments{
\item{prog_train_obj}{A `PrognosiX` object with a trained model}
}
\value{
An S3 object of class `"Prog_Manager"`
}
\description{
Create a deployment manager
}
\examples{
\dontrun{
# Requires a trained PrognosiX object
# mgr <- New_Prog_Manager(prog_obj)
# result <- mgr$prog_predict(new_data)
}
}
