% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub4_nmf.R
\name{nmf_best_rank_analysis}
\alias{nmf_best_rank_analysis}
\title{NMF Best Rank Analysis}
\usage{
nmf_best_rank_analysis(
  data,
  estimate,
  rank_range = 2:4,
  nrun = 10,
  seed = 8891,
  method = "brunet",
  palette_name2 = "Zissou1",
  palette_name1 = "Set3",
  save_dir = here("Subtyping", "cluster_results", "nmf_results")
)
}
\arguments{
\item{data}{Data.}

\item{estimate}{Estimate.}

\item{rank_range}{Rank range.}

\item{nrun}{nrun.}

\item{seed}{Seed.}

\item{method}{Method.}

\item{palette_name2}{Palette.}

\item{palette_name1}{Palette.}

\item{save_dir}{Directory.}
}
\description{
NMF Best Rank Analysis
}
