% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-4.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalize Data for Predictive Modeling}
\usage{
NormalizeData(
  object,
  normalize_method = "min_max_scale",
  group_col = "group",
  max_unique_values = 5,
  alpha = 0.05
)
}
\arguments{
\item{object}{A data object, either of class 'Model_data' or a list containing 'training' and 'testing' data.
The input data should include both training and testing datasets, and the 'Model_data' class should have
associated slots for splitting the data.}

\item{normalize_method}{The normalization method to apply. Default is "min_max_scale". Options include:
"auto", "log_transform", "min_max_scale", "z_score_standardize", etc. This defines how the data will be
normalized, with "min_max_scale" scaling data to a \link{0, 1} range, "z_score_standardize" standardizing data
to have mean 0 and standard deviation 1, etc.}

\item{group_col}{The name of the column containing group labels in the data (e.g., "group"). Default is "group".
This is important for preserving the group label information while normalizing.}

\item{max_unique_values}{Maximum unique values to consider for categorical columns when diagnosing variable types.
Default is 5. This is used to determine if a column should be treated as categorical or continuous during
normalization.}

\item{alpha}{Significance level to determine feature selection thresholds. Default is 0.05. This is used when
applying statistical methods to select relevant features before normalization.}
}
\value{
An updated 'Model_data' object or a list containing the normalized training and testing data,
along with the normalization information. If the input is a 'Model_data' object, the function updates the
object with the normalized data and relevant normalization info. If the input is a list, the function returns
a list containing the normalized datasets and associated information.
}
\description{
This function normalizes the provided training and testing data using the specified normalization method.
It can handle both 'Model_data' objects and lists containing training and testing data. The normalization
process scales the data and returns the normalized data, along with any relevant normalization information.
}
\examples{
# Example using a 'Model_data' object:
model_data <- NormalizeData(object = model_data, normalize_method = "z_score_standardize")

# Example using a list of training and testing data:
data_list <- list(training = training_data, testing = testing_data)
normalized_data <- NormalizeData(object = data_list, normalize_method = "min_max_scale")

# Output includes normalized training and testing data and normalization info.
}
