% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_config.R
\docType{data}
\name{output_config}
\alias{output_config}
\alias{.icare_config}
\title{Global Output Path Configuration for icare Package}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.icare_config
}
\description{
This file provides functions to configure and manage output paths
for all modules (m0-m4). Users can set a custom output directory
that will be used instead of the default code folder paths.
}
\keyword{internal}
