% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_cutsubgroup.R
\name{perform_subgroup_analysis}
\alias{perform_subgroup_analysis}
\title{Perform Subgroup Analysis}
\usage{
perform_subgroup_analysis(
  object,
  status_col = "status",
  id_col = "id",
  time_col = "time",
  max_unique_values = 5,
  methods = "roc",
  seed = 123,
  save_dir = here::here("SurObj", "Data"),
  save_data = TRUE,
  csv_filename = "subgroup_data.csv"
)
}
\arguments{
\item{object}{Object.}

\item{status_col}{Status column.}

\item{id_col}{ID column.}

\item{time_col}{Time column.}

\item{max_unique_values}{Max unique values.}

\item{methods}{Method.}

\item{seed}{Seed.}

\item{save_dir}{Directory.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Perform Subgroup Analysis
}
