% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_09_tsne.R
\name{perform_tsne}
\alias{perform_tsne}
\title{Perform t-SNE}
\usage{
perform_tsne(data, perplexity = 30, dims = 2, iterations = 1000)
}
\arguments{
\item{data}{Data.}

\item{perplexity}{Perplexity.}

\item{dims}{Dims.}

\item{iterations}{Iterations.}
}
\description{
Perform t-SNE
}
\examples{
\dontrun{
  df <- data.frame(x = rnorm(50), y = rnorm(50), z = rnorm(50))
  result <- perform_tsne(df, perplexity = 10, dims = 2)
  plot(result$tsne_df)
}
}
