% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_09_umap.R
\name{perform_umap}
\alias{perform_umap}
\title{Perform UMAP}
\usage{
perform_umap(
  data,
  dims = 2,
  n_neighbors = 15,
  min_dist = 0.1,
  metric = "euclidean"
)
}
\arguments{
\item{data}{Data.}

\item{dims}{Dims.}

\item{n_neighbors}{Neighbors.}

\item{min_dist}{Min dist.}

\item{metric}{Metric.}
}
\description{
Perform UMAP
}
\examples{
\dontrun{
  df <- data.frame(x = rnorm(50), y = rnorm(50), z = rnorm(50))
  result <- perform_umap(df, n_neighbors = 10, dims = 2)
  plot(result$umap_df)
}
}
