% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn16_dca.R
\name{plot_dca_best_model}
\alias{plot_dca_best_model}
\title{Plot DCA for Best Model}
\usage{
plot_dca_best_model(
  object,
  palette_name = "Royal1",
  status_col = "status",
  time_col = "time"
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{palette_name}{Palette.}

\item{status_col}{Status column.}

\item{time_col}{Time column.}
}
\value{
Updated PrognosiX object.
}
\description{
Plots DCA for the best model in PrognosiX object.
}
