% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn16_dca.R
\name{plot_dca_without_save}
\alias{plot_dca_without_save}
\title{Plot DCA Without Save}
\usage{
plot_dca_without_save(
  model_results,
  palette_name = "Royal1",
  type = "continuous",
  status_col = "status",
  time_col = "time"
)
}
\arguments{
\item{model_results}{List of model results.}

\item{palette_name}{Palette name.}

\item{type}{Type (unused).}

\item{status_col}{Status column.}

\item{time_col}{Time column.}
}
\value{
DCA results list.
}
\description{
Plots Decision Curve Analysis without saving.
}
