% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_4_deg.R
\name{plot_deg_violinplot}
\alias{plot_deg_violinplot}
\title{Generate Violin Plot for Differentially Expressed Genes}
\usage{
plot_deg_violinplot(
  last_test_sig,
  data,
  control = "health",
  case = "cancer",
  top_n = 5,
  palette_name = "Royal1",
  name_identity = "deg",
  save_plots = TRUE,
  save_dir = here("StatObject", "deg_info"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  title = "Violin Plot",
  group_col = "group"
)
}
\arguments{
\item{last_test_sig}{A data frame containing the results of differential expression analysis with columns
such as 'change' (e.g., 'Stable', 'Upregulated', 'Downregulated') and 'logFC' (log-fold change).}

\item{data}{A data frame containing the expression data with rows as samples and columns as features.}

\item{control}{The label of the control group (default: "health").}

\item{case}{The label of the case group (default: "cancer").}

\item{top_n}{The number of top significant features to display (default: 5).}

\item{palette_name}{The name of the color palette for the plot.}

\item{name_identity}{The identity for the type of analysis, default is "deg".}

\item{save_plots}{Logical value to indicate whether to save the plot (default: TRUE).}

\item{save_dir}{The directory to save the plot (default: "here('StatObject', 'deg_info')").}

\item{plot_width}{The width of the plot in inches (default: 5).}

\item{plot_height}{The height of the plot in inches (default: 5).}

\item{base_size}{The base font size for the plot (default: 14).}

\item{title}{The title of the plot (default: "Violin Plot").}

\item{group_col}{The column name used to group the data (default: 'group').}
}
\value{
A ggplot object representing the violin plot.
}
\description{
This function generates a violin plot for the top differentially expressed genes based on their log-fold change
and statistical significance. It also calculates the Wilcoxon test p-values for group comparisons and annotates
the plot with significance stars. The plot can be saved to a specified directory.
}
\examples{
# Example usage:
plot_deg_violinplot(last_test_sig = sig_results, data = expression_data)
plot_deg_violinplot(last_test_sig = sig_results, data = expression_data, top_n = 10)
}
