% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_06_deg.R
\name{plot_deg_volcano}
\alias{plot_deg_volcano}
\title{Generate Volcano Plot for Differential Expression}
\usage{
plot_deg_volcano(
  last_test_sig,
  logFC_col = "logFC",
  p_adjust_col = "p.adjust",
  title = "Volcano Plot",
  palette_name = "Zissou1",
  save_dir = NULL,
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)
}
\arguments{
\item{last_test_sig}{A data frame containing the results of differential expression analysis. It should have columns
for log-fold change and adjusted p-values.}

\item{logFC_col}{The name of the column representing log-fold change values (default: "logFC").}

\item{p_adjust_col}{The name of the column representing adjusted p-values (default: "p.adjust").}

\item{title}{The title of the volcano plot (default: "Volcano Plot").}

\item{palette_name}{A string indicating the color palette to use for the plot.}

\item{save_dir}{The directory to save the volcano plot image (default: "here('StatObject', 'deg_info')").}

\item{plot_width}{The width of the plot in inches (default: 5).}

\item{plot_height}{The height of the plot in inches (default: 5).}

\item{base_size}{The base font size for the plot (default: 14).}
}
\value{
A `ggplot` object representing the volcano plot.
}
\description{
This function generates a volcano plot based on the log-fold change and adjusted p-values of differential expression results.
The plot highlights significant and non-significant genes using different colors and includes labels for the significant genes.
The volcano plot is saved as an image in the specified directory.
}
\examples{
if (interactive()) {
  data(iris)
  res <- batch_Wilcoxon(iris[iris$Species != "virginica", ], group_col = "Species")
  plot_deg_volcano(res, save_dir = NULL)
}
}
