% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_rcss.R,
%   R/module_4_pn5_4_rcss_analysis.R
\name{plot_regression_analysis}
\alias{plot_regression_analysis}
\title{Plot Regression Analysis}
\usage{
plot_regression_analysis(
  data,
  method = c("linear", "logistic", "cox"),
  y_label,
  time_col = NULL,
  x_label,
  knots = 3,
  adjust_vars = NULL,
  prob = 0.1,
  save_dir = here::here("PrognosiX", "univariate_analysis")
)

plot_regression_analysis(
  data,
  method = c("linear", "logistic", "cox"),
  y_label,
  time_col = NULL,
  x_label,
  knots = 3,
  adjust_vars = NULL,
  prob = 0.1,
  save_dir = here::here("PrognosiX", "univariate_analysis")
)
}
\arguments{
\item{data}{Data frame.}

\item{method}{Method ("linear", "logistic", "cox").}

\item{y_label}{Outcome variable.}

\item{time_col}{Time variable.}

\item{x_label}{Predictor variable.}

\item{knots}{Number of knots (unused in this function wrapper but kept for signature).}

\item{adjust_vars}{Adjustment variables.}

\item{prob}{Probability threshold.}

\item{save_dir}{Save directory.}
}
\value{
Result from rcssci function.
}
\description{
Plots the results of regression analysis with Restricted Cubic Splines using the rcssci package.
}
