% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_survival_analysis_failed.R,
%   R/module_4_pn5_2_single_var_survival.R
\name{plot_survival_time_distribution}
\alias{plot_survival_time_distribution}
\title{Plot Survival Time Distribution}
\usage{
plot_survival_time_distribution(
  data,
  time_col,
  var_col,
  palette_name = "AsteroidCity1",
  binwidth = 10,
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)

plot_survival_time_distribution(
  data,
  time_col,
  var_col,
  palette_name = "AsteroidCity1",
  binwidth = 10,
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)
}
\arguments{
\item{data}{Data frame.}

\item{time_col}{Time column.}

\item{var_col}{Variable column.}

\item{palette_name}{Palette name.}

\item{binwidth}{Bin width for histogram.}

\item{save_plots}{Logical.}

\item{save_dir}{Save directory.}

\item{plot_width}{Plot width.}

\item{plot_height}{Plot height.}

\item{base_size}{Base font size.}
}
\value{
The ggplot object.
}
\description{
Plots the distribution of survival time for a given variable.
}
