% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_survival_analysis_failed.R,
%   R/module_4_pn5_2_single_var_survival.R
\name{plot_var_kaplan_meier}
\alias{plot_var_kaplan_meier}
\title{Plot Var Kaplan Meier}
\usage{
plot_var_kaplan_meier(
  object,
  time_col = "time",
  status_col = "status",
  var_col = NULL,
  palette_name = "Dark2",
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  use_subgroup_data = FALSE
)

plot_var_kaplan_meier(
  object,
  time_col = "time",
  status_col = "status",
  var_col = NULL,
  palette_name = "Dark2",
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  use_subgroup_data = FALSE
)
}
\arguments{
\item{object}{PrognosiX object or data frame.}

\item{time_col}{Time column name.}

\item{status_col}{Status column name.}

\item{var_col}{Variable column name.}

\item{palette_name}{Palette name.}

\item{save_plots}{Logical.}

\item{save_dir}{Save directory.}

\item{plot_width}{Plot width.}

\item{plot_height}{Plot height.}

\item{base_size}{Base font size.}

\item{use_subgroup_data}{Logical.}

\item{break.x.by}{Break x by.}

\item{pval.method}{Pval method.}

\item{conf.int}{Conf int.}

\item{title}{Title.}

\item{surv.median.line}{Median line.}

\item{risk.table}{Risk table.}

\item{xlab}{Xlab.}

\item{legend.position}{Legend pos.}

\item{legend.title}{Legend title.}
}
\value{
Updated PrognosiX object or plot object.
}
\description{
Wrapper to plot KM curve for a specific variable using PrognosiX object or data frame.
}
