% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_survival_analysis_failed.R,
%   R/module_4_pn5_2_single_var_survival.R
\name{plot_var_survival_time}
\alias{plot_var_survival_time}
\title{Plot Var Survival Time}
\usage{
plot_var_survival_time(
  object,
  time_col = "time",
  status_col = "status",
  var_col = NULL,
  binwidth = 10,
  palette_name = "AsteroidCity1",
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  use_subgroup_data = FALSE
)

plot_var_survival_time(
  object,
  time_col = "time",
  status_col = "status",
  var_col = NULL,
  binwidth = 10,
  palette_name = "AsteroidCity1",
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  use_subgroup_data = FALSE
)
}
\arguments{
\item{object}{PrognosiX object or data frame.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{var_col}{Variable column.}

\item{binwidth}{Bin width.}

\item{palette_name}{Palette name.}

\item{save_plots}{Logical.}

\item{save_dir}{Save directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Base size.}

\item{use_subgroup_data}{Logical.}
}
\value{
Updated PrognosiX object or plot object.
}
\description{
Wrapper function to plot survival time distribution for PrognosiX object or data frame.
}
