% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_06_ensemble.R
\name{PlotModelHeatmap}
\alias{PlotModelHeatmap}
\title{Model Performance Heatmap}
\usage{
PlotModelHeatmap(
  model_obj,
  metrics = c("auc", "Sensitivity", "Specificity", "accuracy_score", "f1_score",
    "Precision"),
  save_plot = FALSE,
  save_dir = NULL
)
}
\arguments{
\item{model_obj}{A trained Train_Model object.}

\item{metrics}{Metrics to display.}

\item{save_plot}{Save? Default FALSE.}

\item{save_dir}{Output directory.}
}
\value{
A ggplot object.
}
\description{
Model Performance Heatmap
}
