% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_06_ensemble.R
\name{PlotModelParallel}
\alias{PlotModelParallel}
\title{Model Performance Parallel Coordinates}
\usage{
PlotModelParallel(
  model_obj,
  top_n = 5,
  metrics = c("auc", "Sensitivity", "Specificity", "accuracy_score", "f1_score"),
  save_plot = FALSE,
  save_dir = NULL
)
}
\arguments{
\item{model_obj}{A trained Train_Model object.}

\item{top_n}{Number of top models. Default 5.}

\item{metrics}{Metrics to display.}

\item{save_plot}{Save? Default FALSE.}

\item{save_dir}{Output directory.}
}
\value{
A ggplot object.
}
\description{
Model Performance Parallel Coordinates
}
