% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_09_nri_idi.R
\name{PlotNRIBars}
\alias{PlotNRIBars}
\title{NRI Reclassification Bar Plot}
\usage{
PlotNRIBars(nri_result, save_plot = FALSE, save_dir = NULL)
}
\arguments{
\item{nri_result}{Output of \code{CalculateCategoryNRI}.}

\item{save_plot}{Logical.}

\item{save_dir}{Output directory.}
}
\value{
A ggplot object.
}
\description{
Shows the number of patients reclassified correctly (green) or
incorrectly (red).
}
