% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_06_ensemble.R
\name{PlotProbDensity}
\alias{PlotProbDensity}
\title{Prediction Probability Density Plot (fixed)}
\usage{
PlotProbDensity(model_obj, data = NULL, save_plot = FALSE, save_dir = NULL)
}
\arguments{
\item{model_obj}{A trained Train_Model object.}

\item{data}{Test data (default: split.data$testing).}

\item{save_plot}{Logical.}

\item{save_dir}{Output directory.}
}
\description{
Prediction Probability Density Plot (fixed)
}
\examples{
\dontrun{
if (interactive()) {
  mtcars$am <- as.factor(mtcars$am)
  model <- CreateModelObject(data = mtcars, group_col = "am")
  set.seed(123)
  idx <- sample(1:nrow(mtcars), 20)
  model@filtered.set <- list(training = mtcars[idx, ], testing = mtcars[-idx, ])
  trained <- ModelTrainAnalysis(model, methods = c("glm"), 
  control = list(method = "cv", number = 3), save_plots = FALSE)
  PlotProbDensity(trained, data = trained@filtered.set$testing, save_plot = FALSE)
  PlotProbStrip(trained, data = trained@filtered.set$testing, save_plot = FALSE)
}
}
}
