% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_06_ensemble.R
\name{PlotProbStrip}
\alias{PlotProbStrip}
\title{Prediction Probability Strip Chart (fixed)}
\usage{
PlotProbStrip(model_obj, data = NULL, save_plot = FALSE, save_dir = NULL)
}
\arguments{
\item{model_obj}{A trained Train_Model object.}

\item{data}{Test data.}

\item{save_plot}{Logical.}

\item{save_dir}{Output directory.}
}
\description{
Prediction Probability Strip Chart (fixed)
}
