% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_08_clinical_analysis.R
\name{PlotThresholdAccuracy}
\alias{PlotThresholdAccuracy}
\title{Accuracy/PPV/NPV vs Threshold Plot with Custom Threshold Markers}
\usage{
PlotThresholdAccuracy(thresh_result, save_plot = FALSE, save_dir = NULL)
}
\arguments{
\item{thresh_result}{Output from \code{CalculateThresholds}.}

\item{save_plot, save_dir}{Output options.}
}
\value{
A ggplot object.
}
\description{
Accuracy/PPV/NPV vs Threshold Plot with Custom Threshold Markers
}
