% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_08_clinical_analysis.R
\name{PlotThresholdConfusion}
\alias{PlotThresholdConfusion}
\title{Confusion Matrix Heatmap (Customizable Colors)}
\usage{
PlotThresholdConfusion(
  thresh_result,
  which_threshold = "Youden",
  save_plot = FALSE,
  save_dir = NULL,
  fill_colors = c("#d8b365", "#f5f5f5", "#5ab4ac")
)
}
\arguments{
\item{thresh_result}{Output from \code{CalculateThresholds}.}

\item{which_threshold}{Name of the threshold.}

\item{save_plot}{Logical. Save plot?}

\item{save_dir}{Output directory.}

\item{fill_colors}{Character vector of colors for gradient.
Default \code{c("#d8b365", "#f5f5f5", "#5ab4ac")}.}
}
\value{
A ggplot object.
}
\description{
Confusion Matrix Heatmap (Customizable Colors)
}
