% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_08_clinical_analysis.R
\name{PlotThresholdWaterfall}
\alias{PlotThresholdWaterfall}
\title{Waterfall Plot for Threshold Classification}
\usage{
PlotThresholdWaterfall(
  thresh_result,
  which_threshold = "Youden",
  save_plot = FALSE,
  save_dir = NULL,
  colors = c("#f1a340", "#998ec3")
)
}
\arguments{
\item{thresh_result}{Output from \code{CalculateThresholds}.}

\item{which_threshold}{Name of the threshold in \code{thresh_result$thresholds}.}

\item{save_plot, save_dir, colors}{Output options.}
}
\value{
A ggplot object.
}
\description{
Waterfall Plot for Threshold Classification
}
