% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_06_hyperparameter_tuning.R
\name{PlotTuningHistory}
\alias{PlotTuningHistory}
\title{Plot Bayesian Optimization History}
\usage{
PlotTuningHistory(model_obj, save_plot = FALSE, save_dir = NULL)
}
\arguments{
\item{model_obj}{A \code{Train_Model} object that has been fine‑tuned.}

\item{save_plot}{Logical.}

\item{save_dir}{Output directory.}
}
\value{
A ggplot.
}
\description{
Shows the best score found so far across iterations.
}
