% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn10_pls_model.R
\name{pls_compute_hr_and_ci}
\alias{pls_compute_hr_and_ci}
\title{Compute HR and CI for PLS (Wrapper)}
\usage{
pls_compute_hr_and_ci(object)
}
\arguments{
\item{object}{PrognosiX object.}
}
\value{
Updated PrognosiX object or results.
}
\description{
Wrapper to compute HR/CI for PLS in PrognosiX.
}
