% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn7_data_filter.R
\name{pn_filtered_set}
\alias{pn_filtered_set}
\title{Get Filtered Dataset}
\usage{
pn_filtered_set(object)
}
\arguments{
\item{object}{PrognosiX object.}
}
\value{
List with 'training' and 'testing'.
}
\description{
Retrieves the filtered training and testing datasets from a PrognosiX object.
}
