% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-2.R
\name{PrepareData}
\alias{PrepareData}
\title{Prepare Data by Converting Factors to Dummy Variables}
\usage{
PrepareData(object = NULL)
}
\arguments{
\item{object}{An object of class 'Model_data' or a data frame. If a 'Model_data' object is provided, the function extracts the data from the 'data.df' slot. If a data frame is provided, it operates directly on the data frame.}
}
\value{
If the input is a 'Model_data' object, the function returns the updated 'Model_data' object with the 'clean.df' slot containing the transformed data. If the input is a data frame, the function returns the transformed data frame.
}
\description{
This function prepares the input data by transforming factor or character variables into binary or dummy variables. The function can handle either a 'Model_data' object or a standard data frame. If the input is a 'Model_data' object, the function updates its \code{clean.df} slot with the transformed data. If the input is a data frame, it directly returns the transformed data.
}
\examples{
\dontrun{
# Example of using with a 'Model_data' object
model_data_obj <- CreateModelObject(data = df)
prepared_model_data <- PrepareData(model_data_obj)

# Example of using with a data frame
df <- data.frame(
  group = factor(c("A", "B", "A", "B")),
  outcome = factor(c("Yes", "No", "Yes", "Yes"))
)
prepared_data <- PrepareData(df)
}
}
