% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn4_data_conversion.R
\name{PrepareDataForPrognosiX}
\alias{PrepareDataForPrognosiX}
\title{Prepare Data For PrognosiX}
\usage{
PrepareDataForPrognosiX(object, group_col = NULL, max_unique_values = 5)
}
\arguments{
\item{object}{PrognosiX object.}

\item{group_col}{Group col.}

\item{max_unique_values}{Max unique values.}
}
\description{
Prepare Data For PrognosiX
}
