% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_08_normalize.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data}
\usage{
preprocess_data(
  data,
  method = "log",
  group_col = "group",
  max_unique_values = 5
)
}
\arguments{
\item{data}{Data frame.}

\item{method}{Preprocessing method.}

\item{group_col}{Group column.}

\item{max_unique_values}{Max unique values.}
}
\value{
Processed data frame.
}
\description{
Preprocess Data
}
\examples{
data(iris)
iris_num <- iris[, 1:4]
proc <- preprocess_data(iris_num, method = "z_score")
head(proc)
}
