% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn7_data_filter.R
\name{Prognos_filter_features}
\alias{Prognos_filter_features}
\title{Filter Features in PrognosiX Object}
\usage{
Prognos_filter_features(
  object,
  time_col = "time",
  status_col = "status",
  use_filtered_features = TRUE
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{use_filtered_features}{Logical.}
}
\value{
Updated PrognosiX object.
}
\description{
Filters the training and testing datasets in a PrognosiX object to retain only selected features.
}
