% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn3_baseline_table.R
\name{Prognos_gaze_analysis}
\alias{Prognos_gaze_analysis}
\title{Prognos Gaze Analysis}
\usage{
Prognos_gaze_analysis(
  object,
  formula = NULL,
  status_col = "status",
  digits = 1,
  show.p = TRUE,
  gaze_method = 3,
  response_var = NULL,
  save_word = TRUE,
  save_dir = here("PrognosiX", "gaze_baseline"),
  use_subgroup_data = FALSE
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{formula}{Formula.}

\item{status_col}{Status col.}

\item{digits}{Digits.}

\item{show.p}{Show P.}

\item{gaze_method}{Method.}

\item{response_var}{Response var.}

\item{save_word}{Save word.}

\item{save_dir}{Save dir.}

\item{use_subgroup_data}{Use subgroup data.}
}
\description{
Prognos Gaze Analysis
}
