% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn2_subgroup_analysis.R
\name{PrognosiX_subgroup_analysis}
\alias{PrognosiX_subgroup_analysis}
\title{Perform Subgroup Analysis for PrognosiX}
\usage{
PrognosiX_subgroup_analysis(
  object,
  status_col = "status",
  id_col = "id",
  time_col = "time",
  max_unique_values = 5,
  methods = "roc",
  seed = 123,
  save_dir = here::here("StatObject", "Data"),
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{status_col}{Status col.}

\item{id_col}{ID col.}

\item{time_col}{Time col.}

\item{max_unique_values}{Max unique values.}

\item{methods}{Method.}

\item{seed}{Seed.}

\item{save_dir}{Save dir.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Perform Subgroup Analysis for PrognosiX
}
