% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn17_extra_val.R
\name{PrognosiXClinicalPrediction}
\alias{PrognosiXClinicalPrediction}
\title{Clinical Prediction with PrognosiX}
\usage{
PrognosiXClinicalPrediction(
  object,
  new_data,
  group_col = "group",
  palette_name = "Royal1",
  save_dir = here::here("PrognosiX", "clinical_predictions"),
  plot_width = 6,
  plot_height = 6,
  alpha = 1,
  base_size = 14
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{new_data}{New data frame.}

\item{group_col}{Group column.}

\item{palette_name}{Palette.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{alpha}{Alpha.}

\item{base_size}{Size.}
}
\value{
List with predictions and plot.
}
\description{
Predicts clinical risk groups using the best model in PrognosiX.
}
