% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn5_4_rcss_analysis.R
\name{Pron_univariate_regression}
\alias{Pron_univariate_regression}
\title{PrognosiX RCS Analysis Wrapper}
\usage{
Pron_univariate_regression(
  object,
  method = "cox",
  y_label,
  time_col,
  x_label,
  knots = 3,
  adjust_vars = NULL,
  prob = 0.1,
  save_dir = here::here("PrognosiX", "univariate_analysis")
)
}
\arguments{
\item{object}{PrognosiX object or data frame.}

\item{method}{Method.}

\item{y_label}{Outcome variable.}

\item{time_col}{Time column.}

\item{x_label}{Predictor variable.}

\item{knots}{Knots.}

\item{adjust_vars}{Adjustment variables.}

\item{prob}{Probability.}

\item{save_dir}{Save directory.}
}
\value{
Updated PrognosiX object or results list.
}
\description{
Wrapper to perform RCS analysis on a PrognosiX object or data frame.
}
