% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_03_remove_missing.R
\name{remove_high_missing}
\alias{remove_high_missing}
\title{Remove Variables and Samples with High Missing Values}
\usage{
remove_high_missing(
  data,
  miss_threshold = 25,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{data}{A data frame.}

\item{miss_threshold}{The threshold percentage of missing values.}

\item{save_dir}{Directory to save cleaned data.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Remove Variables and Samples with High Missing Values
}
\examples{
data(iris)
iris_na <- iris
iris_na[1:5, 1] <- NA
clean <- remove_high_missing(iris_na, miss_threshold = 20, save_data = FALSE)
print(dim(clean$cleaned_data))
}
