% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn14_rsf_model.R
\name{rsf_compute_hr_and_ci}
\alias{rsf_compute_hr_and_ci}
\title{Compute HR and CI for RSF Coefficients}
\usage{
rsf_compute_hr_and_ci(object)
}
\arguments{
\item{object}{PrognosiX object or model.}
}
\value{
Updated PrognosiX object or results.
}
\description{
Computes HR and CI for RSF coefficients (variable importance).
}
