% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_03_feature_sensitivity.R
\name{select_elbow}
\alias{select_elbow}
\title{Select Optimal Feature Count from Elimination Results}
\usage{
select_elbow(
  elim_obj,
  elbow_method = c("difference", "ratio", "perf_tolerance", "stability_window"),
  tol = 0.05,
  window_size = 5,
  stability_tol = 0.01,
  min_features = NULL,
  verbose = TRUE
)
}
\description{
Select Optimal Feature Count from Elimination Results
}
