% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-6.R
\name{SelFeatureSet}
\alias{SelFeatureSet}
\title{Perform Feature Selection and Plot AUC vs. Number of Features}
\usage{
SelFeatureSet(
  object = NULL,
  max_features = NULL,
  xintercept = NULL,
  AUC_change_threshold = 0.01,
  group_col = "group",
  palette_name = "AsteroidCity1",
  data_type = "clean",
  save_plots = TRUE,
  save_dir = here::here("ModelData", "sel_feature"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  seed = 123
)
}
\arguments{
\item{object}{An object of type \code{'Model_data'} or a list containing 'features' and 'label'. If 'Model_data'
is provided, the function will extract the training data for feature selection.}

\item{max_features}{An integer indicating the maximum number of features to select. Defaults to the number of columns
in the feature matrix.}

\item{xintercept}{A numeric value indicating where to place a vertical line in the plot to mark the optimal number of features.}

\item{AUC_change_threshold}{A numeric threshold to detect change points in the AUC values. Default is 0.01.}

\item{group_col}{The name of the response column (group label) in the dataset. Default is "group".}

\item{palette_name}{Character string specifying the palette name for the plot colors. Default is "AsteroidCity1".}

\item{data_type}{Character string indicating the type of data to use, either 'clean' or 'scale'. Default is "clean".}

\item{save_plots}{Logical value indicating whether to save the generated plot as a PDF. Default is \code{TRUE}.}

\item{save_dir}{The directory where the plot should be saved. Defaults to \code{here::here("ModelData", "sel_feature")}.}

\item{plot_width}{Numeric value specifying the width of the output plot in inches. Default is 5.}

\item{plot_height}{Numeric value specifying the height of the output plot in inches. Default is 5.}

\item{base_size}{Numeric value specifying the base font size for the plot. Default is 14.}

\item{seed}{A numeric value for setting the random seed for reproducibility. Default is 123.}
}
\value{
A ggplot object showing the AUC vs. the number of features for each method. If the input is a 'Model_data'
object, the updated object with feature selection results is returned.
}
\description{
This function performs feature selection using three different methods (IV, MIC, MI), calculates the corresponding
AUC values, and identifies the best feature set based on AUC changes. The function can optionally save the plot
showing AUC vs. the number of selected features for each method. The function also supports updating a 'Model_data'
object with the feature selection results.
}
\examples{
\dontrun{
library(ggplot2)
library(wesanderson)
library(ggprism)
example_data <- list(features = matrix(rnorm(100), ncol = 10), label = sample(1:2, 10, replace = TRUE))
result <- SelFeatureSet(example_data, max_features = 8)
}
}
