% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_03_deploy.R
\name{set_prog_app_text}
\alias{set_prog_app_text}
\title{Set custom text for the Prognosis Terminal}
\usage{
set_prog_app_text(...)
}
\arguments{
\item{...}{Named arguments for text keys (see default list)}
}
\description{
Set custom text for the Prognosis Terminal
}
\examples{
\dontrun{
set_prog_app_text(title = "My Prognosis App", prediction_portal = "Risk Calculator")
}
}
