% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_07_evaluate.R
\name{silhouette_score}
\alias{silhouette_score}
\title{Silhouette Score}
\usage{
silhouette_score(clustered.data, label_col = "group")
}
\arguments{
\item{clustered.data}{Clustered data.}

\item{label_col}{Column for label.}
}
\value{
Numeric average silhouette score.
}
\description{
Silhouette Score
}
\examples{
\dontrun{
  df <- data.frame(x = rnorm(30), y = rnorm(30), group = sample(1:3, 30, replace = TRUE))
  sil <- silhouette_score(df)
  cat("Silhouette Score:", sil, "\n")
}
}
