% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{SplitByMeta}
\alias{SplitByMeta}
\title{Split an object into a list by a metadata column}
\usage{
SplitByMeta(object, col)
}
\arguments{
\item{object}{A \code{Stat}, \code{Subtyping}, or \code{PrognosiX} object.}

\item{col}{Column name in \code{info.data}.}
}
\value{
A named list of objects, one element per unique level of \code{col}.
}
\description{
Splits the object into a named list of objects, one per unique value in the
specified \code{info.data} column.  Equivalent to Seurat's
\code{SplitObject}.
}
\examples{
\dontrun{
split_list <- SplitByMeta(stat_obj_test, col = "SWAB")
}
}
