% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn6_split_dataset.R
\name{SplitData}
\alias{SplitData}
\title{Split Data into Training and Testing Sets}
\usage{
SplitData(
  data,
  unwanted_var = NULL,
  train_ratio = 0.7,
  test_ratio = 0.3,
  seed = 123
)
}
\arguments{
\item{data}{Data frame.}

\item{unwanted_var}{Vector of unwanted variable names to remove.}

\item{train_ratio}{Training set ratio (default 0.7).}

\item{test_ratio}{Testing set ratio (default 0.3).}

\item{seed}{Random seed.}
}
\value{
A list containing 'training' and 'testing' data frames.
}
\description{
Splits a data frame into training and testing sets based on a ratio.
}
