% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn6_split_dataset.R
\name{SplitDataPrognosiX}
\alias{SplitDataPrognosiX}
\title{Split Data for PrognosiX Object}
\usage{
SplitDataPrognosiX(
  object = NULL,
  unwanted_var = NULL,
  train_ratio = 0.7,
  test_ratio = 0.3,
  seed = 123,
  use_subgroup_data = FALSE
)
}
\arguments{
\item{object}{PrognosiX object or data frame.}

\item{unwanted_var}{Unwanted variables.}

\item{train_ratio}{Training ratio.}

\item{test_ratio}{Testing ratio.}

\item{seed}{Seed.}

\item{use_subgroup_data}{Logical.}
}
\value{
Updated PrognosiX object or list of split data.
}
\description{
Splits data within a PrognosiX object into training and testing sets.
}
