% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_01_object.R
\docType{class}
\name{Stat-class}
\alias{Stat-class}
\alias{Stat}
\title{Class to store statistical analysis results}
\value{
An object of class 'Stat' containing the slots mentioned above.
}
\description{
This class is designed to store and manage various components of statistical analysis.
It includes raw and cleaned data, processed results, descriptive statistics, correlation results,
variable types, and metadata related to the analysis.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{A data.frame containing the raw input data.}

\item{\code{clean.data}}{A data.frame containing the cleaned data after preprocessing.}

\item{\code{info.data}}{A data.frame containing additional information related to the data.}

\item{\code{scale.data}}{A data.frame containing the scaled version of the data.}

\item{\code{meta.featurename}}{A character vector containing metadata feature names.}

\item{\code{group_col}}{The grouping column in the data, which can be of any type.}

\item{\code{baseline.table}}{A table that stores baseline information, can be of any type.}

\item{\code{process.info}}{A list containing information about the processing steps.}

\item{\code{compute.descriptive}}{A list containing results of descriptive statistics computations.}

\item{\code{corr.result}}{A list containing the results of correlation analysis.}

\item{\code{var.result}}{A list containing the results of variance analysis.}

\item{\code{variable.types}}{A list specifying the types of the variables in the dataset.}
}}

