% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_04_descriptive.R
\name{stat_convert_variables}
\alias{stat_convert_variables}
\title{Convert Variables in a 'Stat' Object or Data Frame}
\usage{
stat_convert_variables(
  object,
  group_col = "group",
  max_unique_values = 5,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{object}{An object of class \code{Stat} or a data frame. If a \code{Stat}
object is provided, the function will update its \code{clean.data} slot.}

\item{group_col}{A string representing the column name that groups the data.
Default is \code{"group"}. This column is used to determine the type of variables.}

\item{max_unique_values}{The maximum number of unique values allowed for a
variable to be considered as numeric. Default is \code{5}.}

\item{save_dir}{A character string specifying the directory where the converted
data will be saved as a CSV file. If \code{NULL}, the default output directory
is used. Default is \code{NULL}.}

\item{save_data}{Logical. If \code{TRUE}, saves the converted data to a CSV file.
Default is \code{TRUE}.}

\item{csv_filename}{A character string specifying the name of the CSV file to
save the converted data. Default is \code{"clean_data.csv"}.}
}
\value{
If the input is a \code{Stat} object, returns the updated \code{Stat}
  object with the converted data stored in the \code{clean.data} slot.
  If the input is a data frame, returns the converted data frame.
}
\description{
This function converts variables in a given object (either of class 'Stat' or a
data frame) to numeric or factor types based on the information provided by
\code{\link{diagnose_variable_type}}. If the input is a 'Stat' object, the
conversion will update its \code{clean.data} slot. If the input is a data frame,
it will directly return the converted data frame.
}
\examples{
stat <- CreateStatObject(raw.data = mtcars, group_col = "cyl")
stat <- stat_convert_variables(stat, save_data = FALSE)
str(stat@clean.data)
}
\seealso{
\code{\link{diagnose_variable_type}}, \code{\link{convert_variables}}
}
