% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-3.R
\name{stat_correlated_features}
\alias{stat_correlated_features}
\title{Identify Highly Correlated Features from a 'Stat' Object or Data Frame}
\usage{
stat_correlated_features(
  object,
  correlation_threshold = 0.95,
  palette_name = "Zissou1",
  data_type = "scale",
  save_plots = TRUE,
  save_dir = here("StatObject", "correlation_info"),
  save_data = TRUE,
  csv_filename = "correlation_heatmap_data.csv"
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame containing the features to be analyzed.}

\item{correlation_threshold}{A numeric value specifying the threshold for correlation. Features with a correlation
greater than this value will be considered correlated (default is 0.95).}

\item{palette_name}{A string specifying the color palette to be used for the heatmap}

\item{data_type}{A string indicating which type of data to extract from the 'Stat' object. Options are "clean" or "scale" (default is "scale").}

\item{save_dir}{The directory where the heatmap plot will be saved, if \code{save_plots} is TRUE (default is \code{here("StatObject", "correlation_info")}).}
}
\value{
If the input is a 'Stat' object, the function returns the updated 'Stat' object with the correlation results added to the 'corr.result' slot.
If the input is a data frame, the function returns the results of the correlation analysis, including the correlation matrix,
identified correlated features, and other relevant data.
}
\description{
This function performs correlation analysis on a given data frame or 'Stat' object, identifying
features that are highly correlated based on a specified threshold. It returns the result of
the correlation analysis and optionally updates the 'Stat' object with the correlation results.
}
\examples{
# Assuming 'stat_object' is an instance of 'Stat' class
result <- stat_correlated_features(stat_object, correlation_threshold = 0.9)

# If input is a data frame
result <- stat_correlated_features(my_data_frame, correlation_threshold = 0.85)

}
