% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_05_handle_missing.R
\name{stat_miss_processed}
\alias{stat_miss_processed}
\title{Process Missing Values for Stat Object or Data Frame}
\usage{
stat_miss_processed(
  object,
  m = 5,
  impute_method = "mice",
  group_col = "group",
  miss_threshold = 20,
  max_unique_values = 5,
  return_imputation_info = TRUE,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{object}{Stat object or data frame.}

\item{m}{MICE imputations.}

\item{impute_method}{Method.}

\item{group_col}{Group column.}

\item{miss_threshold}{Missing threshold.}

\item{max_unique_values}{Max unique values.}

\item{return_imputation_info}{Logical.}

\item{save_dir}{Save directory.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Process Missing Values for Stat Object or Data Frame
}
\examples{
stat <- CreateStatObject(raw.data = mtcars, group_col = "cyl")
stat <- stat_miss_processed(stat, impute_method = "median_mode", return_imputation_info = FALSE)
head(stat@clean.data)
}
