% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_08_normalize.R
\name{stat_normalize_process}
\alias{stat_normalize_process}
\title{Normalize Data in Stat Object}
\usage{
stat_normalize_process(
  object,
  method = "auto",
  group_col = "group",
  max_unique_values = 5,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "scale_data.csv"
)
}
\arguments{
\item{object}{Stat object.}

\item{method}{Normalization method.}

\item{group_col}{Group column.}

\item{max_unique_values}{Max unique values.}

\item{save_dir}{Save directory.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Normalize Data in Stat Object
}
\examples{
stat <- CreateStatObject(clean.data = mtcars, group_col = "cyl")
stat <- stat_normalize_process(stat, method = "z_score")
head(stat@scale.data)
}
