% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub4_nmf.R
\name{Sub_best_rank_analysis}
\alias{Sub_best_rank_analysis}
\title{Sub Best Rank Analysis}
\usage{
Sub_best_rank_analysis(
  object,
  nrun = 10,
  seed = 8891,
  method = "brunet",
  palette_name2 = "Zissou1",
  palette_name1 = "Set3",
  save_dir = here("Subtyping", "cluster_results", "nmf_results")
)
}
\arguments{
\item{object}{Subtyping object.}

\item{nrun}{nrun.}

\item{seed}{Seed.}

\item{method}{Method.}

\item{palette_name2}{Palette.}

\item{palette_name1}{Palette.}

\item{save_dir}{Directory.}
}
\description{
Sub Best Rank Analysis
}
