% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_kmeans_alt_2.R
\name{Sub_kmeans_with_optimal_k_alt}
\alias{Sub_kmeans_with_optimal_k_alt}
\title{Sub K-means with Optimal K (Alternative)}
\usage{
Sub_kmeans_with_optimal_k_alt(
  object,
  use_scaled_data = TRUE,
  palette_name = "Zissou1",
  save_plots = TRUE,
  save_dir = here("Subtyping", "cluster_results", "kmeans_result"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  nstart = 50,
  seed = 123,
  k.max = 15
)
}
\arguments{
\item{object}{Subtyping object.}

\item{use_scaled_data}{Logical.}

\item{palette_name}{Palette.}

\item{save_plots}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Base size.}

\item{nstart}{nstart.}

\item{seed}{Seed.}

\item{k.max}{K max.}
}
\description{
Sub K-means with Optimal K (Alternative)
}
