% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_04_nmf.R
\name{Sub_nmf_plot_eval}
\alias{Sub_nmf_plot_eval}
\title{Plot NMF rank evaluation using NMF native graphics}
\usage{
Sub_nmf_plot_eval(object, save_dir = NULL, width = 15, height = 12)
}
\arguments{
\item{object}{Subtyping object.}

\item{save_dir}{Output directory.}

\item{width}{PDF width.}

\item{height}{PDF height.}
}
\description{
Plot NMF rank evaluation using NMF native graphics
}
\examples{
\dontrun{
  # Assuming 'obj' is a Subtyping object with nmf.result
  Sub_nmf_plot_eval(obj, save_dir = "./nmf_results")
}
}
