% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_create_survival_time.R
\name{Sub_survival_data}
\alias{Sub_survival_data}
\title{Sub Survival Data}
\usage{
Sub_survival_data(
  object,
  time_col = "time",
  status_col = "status",
  keep_all = FALSE
)
}
\arguments{
\item{object}{Subtyping object.}

\item{time_col}{Time col.}

\item{status_col}{Status col.}

\item{keep_all}{Keep all.}
}
\description{
Sub Survival Data
}
