% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_09_tsne.R
\name{Sub_tsne_analyse}
\alias{Sub_tsne_analyse}
\title{Sub t-SNE Analyse}
\usage{
Sub_tsne_analyse(
  object,
  perplexity = 30,
  dims = 2,
  iterations = 1000,
  use_scaled_data = TRUE,
  seed = 123
)
}
\arguments{
\item{object}{Subtyping object.}

\item{perplexity}{Perplexity.}

\item{dims}{Dims.}

\item{iterations}{Iterations.}

\item{use_scaled_data}{Logical.}

\item{seed}{Numeric.}
}
\description{
Sub t-SNE Analyse
}
\examples{
\dontrun{
  # Assuming 'obj' is a Subtyping object
  obj <- Sub_tsne_analyse(obj, perplexity = 20, dims = 2)
}
}
