% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_09_umap.R
\name{Sub_umap_analyse}
\alias{Sub_umap_analyse}
\title{Sub UMAP Analyse}
\usage{
Sub_umap_analyse(
  object,
  dims = 2,
  n_neighbors = 15,
  min_dist = 0.1,
  metric = "euclidean",
  use_scaled_data = TRUE,
  seed = 123
)
}
\arguments{
\item{object}{Subtyping object.}

\item{dims}{Dims.}

\item{n_neighbors}{Neighbors.}

\item{min_dist}{Min dist.}

\item{metric}{Metric.}

\item{use_scaled_data}{Logical.}

\item{seed}{Numeric.}
}
\description{
Sub UMAP Analyse
}
\examples{
\dontrun{
  # Assuming 'obj' is a Subtyping object
  obj <- Sub_umap_analyse(obj, n_neighbors = 10, dims = 2)
}
}
