% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_01_object.R
\docType{class}
\name{Subtyping-class}
\alias{Subtyping-class}
\alias{Subtyping}
\title{Subtyping Class}
\description{
Subtyping Class
}
\section{Slots}{

\describe{
\item{\code{clean.data}}{Data frame.}

\item{\code{info.data}}{Data frame.}

\item{\code{scale.data}}{Data frame.}

\item{\code{Optimal.cluster}}{Any.}

\item{\code{cluster.results}}{Any.}

\item{\code{visualization.results}}{Any.}

\item{\code{clustered.data}}{Data frame.}

\item{\code{evaluation_results}}{List.}
}}

\examples{
\dontrun{
  obj <- new("Subtyping",
             clean.data = data.frame(x = 1:3, y = 4:6),
             info.data = data.frame(id = letters[1:3]),
             scale.data = data.frame(x = c(0, 0.5, 1), y = c(0, 0.5, 1)))
}
}
