% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_univariate_analysis.R
\name{Sur_cox_univariate_analysis}
\alias{Sur_cox_univariate_analysis}
\title{Sur Cox Univariate Analysis}
\usage{
Sur_cox_univariate_analysis(
  object,
  formula = NULL,
  status_col = "status",
  time_col = "time",
  selected_vars = NULL,
  covariates = NULL,
  P_value = 0.05,
  save_dir = here("SurObj", "univariate_analysis")
)
}
\arguments{
\item{object}{SurObj.}

\item{formula}{Formula.}

\item{status_col}{Status col.}

\item{time_col}{Time col.}

\item{selected_vars}{Selected vars.}

\item{covariates}{Covariates.}

\item{P_value}{P value.}

\item{save_dir}{Save dir.}
}
\description{
Sur Cox Univariate Analysis
}
