% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_create_baseline_table.R
\name{Sur_gaze_analysis}
\alias{Sur_gaze_analysis}
\title{Sur Gaze Analysis}
\usage{
Sur_gaze_analysis(
  object,
  formula = NULL,
  status_col = "status",
  digits = 1,
  show.p = TRUE,
  gaze_method = 3,
  response_var = "group",
  save_word = TRUE,
  use_subgroup_data = FALSE,
  save_dir = here("SurObj", "gaze_baseline")
)
}
\arguments{
\item{object}{SurObj.}

\item{formula}{Formula.}

\item{status_col}{Status column.}

\item{digits}{Digits.}

\item{show.p}{Show P.}

\item{gaze_method}{Method.}

\item{response_var}{Response var.}

\item{save_word}{Save word.}

\item{use_subgroup_data}{Use subgroup data.}

\item{save_dir}{Save dir.}
}
\description{
Sur Gaze Analysis
}
