% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_rcss.R
\name{Sur_univariate_regression}
\alias{Sur_univariate_regression}
\title{Sur Univariate Regression}
\usage{
Sur_univariate_regression(
  object,
  method = "cox",
  y_label,
  time_col = "time",
  x_label,
  knots = 3,
  adjust_vars = NULL,
  prob = 0.1,
  save_dir = here("SurObj", "univariate_analysis")
)
}
\arguments{
\item{object}{SurObj.}

\item{method}{Method.}

\item{y_label}{Y label.}

\item{time_col}{Time col.}

\item{x_label}{X label.}

\item{knots}{Knots.}

\item{adjust_vars}{Adjust vars.}

\item{prob}{Prob.}

\item{save_dir}{Save dir.}
}
\description{
Sur Univariate Regression
}
