% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_create_surv_obj.R
\docType{class}
\name{SurObj-class}
\alias{SurObj-class}
\alias{SurObj}
\title{SurObj Class}
\description{
SurObj Class
}
\section{Slots}{

\describe{
\item{\code{clean.data}}{Data frame.}

\item{\code{info.data}}{Data frame.}

\item{\code{clustered.data}}{Data frame.}

\item{\code{survival.data}}{Data frame.}

\item{\code{sub.data}}{Data frame.}

\item{\code{baseline.table}}{Any.}

\item{\code{univariate.analysis}}{List.}

\item{\code{survival.analysis}}{List.}

\item{\code{rcss.results}}{Any.}
}}

