% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_02_framework.R
\name{surv_plot_comparison_auc}
\alias{surv_plot_comparison_auc}
\title{Plot Comparison Time-dependent AUC (Train vs Validation)}
\usage{
surv_plot_comparison_auc(learner, train_task, val_task)
}
\arguments{
\item{learner}{Trained learner}

\item{train_task}{Training task}

\item{val_task}{Validation task}
}
\value{
ggplot object
}
\description{
Plot Comparison Time-dependent AUC (Train vs Validation)
}
