% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_survival_analysis_failed.R,
%   R/module_4_pn5_2_single_var_survival.R
\name{survival_analysis_var_plot}
\alias{survival_analysis_var_plot}
\title{Survival Analysis Var Plot}
\usage{
survival_analysis_var_plot(
  data,
  time_col = "time",
  status_col = "status",
  var_col = NULL,
  palette_name = "Dark2",
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)

survival_analysis_var_plot(
  data,
  time_col = "time",
  status_col = "status",
  var_col = NULL,
  palette_name = "Dark2",
  save_plots = TRUE,
  save_dir = here::here("PrognosiX", "univariate_analysis"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)
}
\arguments{
\item{data}{Data frame.}

\item{time_col}{Time column name.}

\item{status_col}{Status column name.}

\item{var_col}{Variable to stratify by.}

\item{palette_name}{Palette name.}

\item{save_plots}{Logical.}

\item{save_dir}{Save directory.}

\item{plot_width}{Plot width.}

\item{plot_height}{Plot height.}

\item{base_size}{Base font size.}

\item{break.x.by}{Break x by.}

\item{pval.method}{Pval method.}

\item{conf.int}{Conf int.}

\item{title}{Title.}

\item{surv.median.line}{Median line.}

\item{risk.table}{Risk table.}

\item{xlab}{Xlab.}

\item{legend.position}{Legend pos.}

\item{legend.title}{Legend title.}
}
\value{
A list containing the KM object and the combined plot.
}
\description{
Performs survival analysis and generates Kaplan-Meier plots for a specific variable.
}
