% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn15_all_models.R
\name{train_all_models}
\alias{train_all_models}
\title{Train All Models}
\usage{
train_all_models(
  object,
  model_list = c("ridge", "lasso", "pls", "coxboost", "coxph", "superpc", "rsf")
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{model_list}{List of models to train ("ridge", "lasso", "pls", "coxboost", "coxph", "superpc", "rsf").}
}
\value{
Updated PrognosiX object.
}
\description{
Trains multiple survival models specified in the list.
}
