% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn11_coxboost_model.R
\name{train_coxboost_base}
\alias{train_coxboost_base}
\title{Train CoxBoost Model (Base)}
\usage{
train_coxboost_base(
  train_data,
  time_col = "time",
  status_col = "status",
  penalty = 100,
  seed = 123
)
}
\arguments{
\item{train_data}{Training data.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{penalty}{Penalty.}

\item{seed}{Seed.}
}
\value{
List containing model.
}
\description{
Trains a CoxBoost model (base function).
}
