% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn11_coxboost_model.R
\name{train_coxboost_model}
\alias{train_coxboost_model}
\title{Train CoxBoost Model for PrognosiX}
\usage{
train_coxboost_model(
  object,
  time_col = "time",
  status_col = "status",
  penalty = 100,
  seed = 123
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{penalty}{Penalty.}

\item{seed}{Seed.}
}
\value{
Updated PrognosiX object or model list.
}
\description{
Wrapper to train CoxBoost model on PrognosiX object.
}
