% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn12_coxph_model.R
\name{train_coxph_base}
\alias{train_coxph_base}
\title{Train CoxPH Model (Base)}
\usage{
train_coxph_base(
  train_data,
  time_col = "time",
  status_col = "status",
  seed = 123
)
}
\arguments{
\item{train_data}{Training data.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{seed}{Seed.}
}
\value{
List containing model.
}
\description{
Trains a Cox Proportional Hazards (CoxPH) model (base function).
}
