% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn12_coxph_model.R
\name{train_coxph_model}
\alias{train_coxph_model}
\title{Train CoxPH Model for PrognosiX}
\usage{
train_coxph_model(object, time_col = "time", status_col = "status", seed = 123)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{seed}{Seed.}
}
\value{
Updated PrognosiX object or model list.
}
\description{
Wrapper to train CoxPH model on PrognosiX object.
}
