% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn8_lasso_model.R
\name{train_lasso_base}
\alias{train_lasso_base}
\title{Train LASSO Model (Base)}
\usage{
train_lasso_base(
  train_data,
  time_col = "time",
  status_col = "status",
  nfolds = 10,
  palette_name = "AsteroidCity1",
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "lasso_model"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  seed = 123
)
}
\arguments{
\item{train_data}{Training data.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{nfolds}{Number of folds for CV.}

\item{palette_name}{Palette name.}

\item{save_plot}{Logical.}

\item{save_dir}{Save directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Base font size.}

\item{seed}{Seed.}
}
\value{
List containing model, lambda, cv_fit, and plot.
}
\description{
Trains a LASSO Cox model (base function).
}
