% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn8_lasso_model.R
\name{train_lasso_model}
\alias{train_lasso_model}
\title{Train LASSO Model for PrognosiX}
\usage{
train_lasso_model(
  object,
  time_col = "time",
  status_col = "status",
  nfolds = 10,
  palette_name = "AsteroidCity1",
  save_plot = TRUE,
  save_dir = here::here("PrognosiX", "lasso_model"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  seed = 123
)
}
\arguments{
\item{object}{PrognosiX object or list.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{nfolds}{Folds.}

\item{palette_name}{Palette.}

\item{save_plot}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Size.}

\item{seed}{Seed.}
}
\value{
Updated PrognosiX object or model list.
}
\description{
Wrapper to train LASSO model on PrognosiX object.
}
