% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_01_object.R
\docType{class}
\name{Train_Model-class}
\alias{Train_Model-class}
\alias{Train_Model}
\title{Train_Model S4 Class}
\description{
An S4 class to store cleaned data, grouping information, and various results from the
modeling pipeline, including feature selection, model training.
}
\section{Slots}{

\describe{
\item{\code{data.df}}{A \code{data.frame} containing the original input data.}

\item{\code{clean.df}}{A \code{data.frame} containing cleaned and preprocessed data.}

\item{\code{group_col}}{A column name or identifier indicating group labels.}

\item{\code{process.info}}{A \code{list} storing information about class balancing.}

\item{\code{split.data}}{A \code{list} of training/testing data splits.}

\item{\code{split.scale.data}}{A \code{list} of scaled training/testing data.}

\item{\code{feature.selection}}{A \code{list} of selected features for modeling.}

\item{\code{feature.result}}{A \code{list} containing feature evaluation results.}

\item{\code{filtered.set}}{A \code{list} of data or variables filtered by some criteria.}

\item{\code{train.models}}{A \code{list} of trained models.}

\item{\code{all.results}}{A \code{list} storing evaluation results from all models.}

\item{\code{best.model.result}}{A \code{list} containing the best model and its performance.}
}}

