% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn10_pls_model.R
\name{train_pls_model}
\alias{train_pls_model}
\title{Train PLS Model for PrognosiX}
\usage{
train_pls_model(
  object,
  time_col = "time",
  status_col = "status",
  nfolds = 10,
  method = "efron",
  plot.it = TRUE,
  se = TRUE,
  scaleX = TRUE
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{nfolds}{Folds.}

\item{method}{Method.}

\item{plot.it}{Plot CV.}

\item{se}{Standard errors.}

\item{scaleX}{Scale predictors.}
}
\value{
Updated PrognosiX object or model list.
}
\description{
Wrapper to train PLS model on PrognosiX object.
}
