% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn14_rsf_model.R
\name{train_rsf_base}
\alias{train_rsf_base}
\title{Train RSF Model (Base)}
\usage{
train_rsf_base(
  train_data,
  time_col = "time",
  status_col = "status",
  ntree = 1000,
  nodesize = 5,
  seed = 123
)
}
\arguments{
\item{train_data}{Training data.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{ntree}{Number of trees.}

\item{nodesize}{Node size.}

\item{seed}{Seed.}
}
\value{
List containing model.
}
\description{
Trains a Random Survival Forest (RSF) model (base function).
}
