% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn14_rsf_model.R
\name{train_rsf_model}
\alias{train_rsf_model}
\title{Train RSF Model for PrognosiX}
\usage{
train_rsf_model(
  object,
  time_col = "time",
  status_col = "status",
  ntree = 1000,
  nodesize = 5,
  seed = 123
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{ntree}{Number of trees.}

\item{nodesize}{Node size.}

\item{seed}{Seed.}
}
\value{
Updated PrognosiX object or model list.
}
\description{
Wrapper to train RSF model on PrognosiX object.
}
