% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn13_superpc_model.R
\name{train_superpc_base}
\alias{train_superpc_base}
\title{Train SuperPC Model (Base)}
\usage{
train_superpc_base(
  train_data,
  time_col = "time",
  status_col = "status",
  seed = 123
)
}
\arguments{
\item{train_data}{Training data.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{seed}{Seed.}
}
\value{
List containing model and threshold.
}
\description{
Trains a SuperPC model (base function).
}
