% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn13_superpc_model.R
\name{train_superpc_model}
\alias{train_superpc_model}
\title{Train SuperPC Model for PrognosiX}
\usage{
train_superpc_model(
  object,
  time_col = "time",
  status_col = "status",
  seed = 123
)
}
\arguments{
\item{object}{PrognosiX object.}

\item{time_col}{Time column.}

\item{status_col}{Status column.}

\item{seed}{Seed.}
}
\value{
Updated PrognosiX object or model list.
}
\description{
Wrapper to train SuperPC model on PrognosiX object.
}
