% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_04_Extensions.R
\name{tune_with_bayes}
\alias{tune_with_bayes}
\title{Tune a survival learner using Bayesian optimisation (robust fallback)}
\usage{
tune_with_bayes(object, learner_id, tuning_budget = 30, ...)
}
\arguments{
\item{object}{A `TaskSurv` or `PrognosiX` object.}

\item{learner_id}{Character; learner to tune.}

\item{tuning_budget}{Number of evaluations.}

\item{...}{Additional arguments passed to `mlr3tuning::tnr("mbo")`.}
}
\value{
Same structure as `surv_train_and_tune`.
}
\description{
Tune a survival learner using Bayesian optimisation (robust fallback)
}
