% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_03_deploy.R
\name{use_app_theme_dark}
\alias{use_app_theme_dark}
\title{Predefined theme: dark (original dark background)}
\usage{
use_app_theme_dark()
}
\description{
Predefined theme: dark (original dark background)
}
\examples{
\dontrun{
use_app_theme_dark()
}
}
