% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_03_deploy.R
\name{use_app_theme_grey}
\alias{use_app_theme_grey}
\title{Predefined theme: advanced grey (light background, dark text)}
\usage{
use_app_theme_grey()
}
\description{
Predefined theme: advanced grey (light background, dark text)
}
\examples{
\dontrun{
use_app_theme_grey()
}
}
