% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_03_deploy.R
\name{use_app_theme_light}
\alias{use_app_theme_light}
\title{Predefined theme: light (clean white, blue accents)}
\usage{
use_app_theme_light()
}
\description{
Predefined theme: light (clean white, blue accents)
}
\examples{
\dontrun{
use_app_theme_light()
}
}
