% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_4_deg.R
\name{VarFeature_violinplot}
\alias{VarFeature_violinplot}
\title{Generate Violin Plot for Differential Features}
\usage{
VarFeature_violinplot(
  object,
  control = "health",
  case = "cancer",
  top_n = 5,
  palette_name = "Royal1",
  name_identity = "deg",
  data_type = "clean",
  save_dir = here("StatObject", "deg_info"),
  plot_width = 5,
  plot_height = 5,
  base_size = 10
)
}
\arguments{
\item{object}{The input object, which can either be of class 'Stat' or a data frame.}

\item{control}{The label for the control group in the data (default is 'health').}

\item{case}{The label for the case group in the data (default is 'cancer').}

\item{top_n}{The number of top features to plot based on differential expression (default is 5).}

\item{palette_name}{The color palette to use for the plot.}

\item{name_identity}{The identifier for the features, used in plot titles or labels (default is 'deg').}

\item{data_type}{Specifies whether to use "clean" or "scaled" data (default is "clean").}

\item{save_dir}{The directory where the plot will be saved (default is \code{here("StatObject", "deg_info")}).}

\item{plot_width}{The width of the plot in centimeters (default is 5).}

\item{plot_height}{The height of the plot in centimeters (default is 5).}

\item{base_size}{The base font size for the plot (default is 14).}
}
\value{
If the input is a 'Stat' object, the function returns the updated 'Stat' object with the violin plot stored in the 'var.result' slot. If the input is a data frame, the function returns the generated violin plot.
}
\description{
This function generates a violin plot to visualize the distribution of the top differentially expressed (DE) features between two groups (e.g., control vs case) based on statistical significance. The function allows users to customize the number of features, color palette, and save the plot to a specified directory.
}
\examples{
# Example usage with a 'Stat' object:
var_plot <- VarFeature_violinplot(object = stat_object, control = 'healthy', case = 'cancer', top_n = 10)

# Example usage with a data frame:
var_plot_df <- VarFeature_violinplot(object = df, control = 'healthy', case = 'cancer', top_n = 5)

}
