% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-8-balance.R
\name{visualize_balance}
\alias{visualize_balance}
\title{Visualize the balance of the dataset before and after balancing.}
\usage{
visualize_balance(
  original_data,
  balanced_data,
  group_col = "group",
  palette_name = "Royal1",
  base_size = 14,
  save_plots = TRUE,
  save_dir = here("ModelData", "balacing_info"),
  plot_width = 5,
  plot_height = 5
)
}
\arguments{
\item{original_data}{A data frame containing the original (imbalanced) dataset.}

\item{balanced_data}{A data frame containing the balanced dataset.}

\item{group_col}{A string specifying the column name that represents the class labels in both datasets. Default is "group".}

\item{palette_name}{A string specifying the name of the color palette for the plot. Default is "Royal1".}

\item{base_size}{An integer specifying the base font size for the plot. Default is 14.}

\item{save_plots}{A logical value indicating whether to save the plot as a PDF. Default is \code{TRUE}.}

\item{save_dir}{A string specifying the directory where the plot will be saved. Default is "ModelData/balacing_info" using \code{here()}.}

\item{plot_width}{A numeric value specifying the width of the plot when saved. Default is 5.}

\item{plot_height}{A numeric value specifying the height of the plot when saved. Default is 5.}
}
\value{
A ggplot object representing the class distribution before and after balancing.
}
\description{
This function generates a bar plot to visualize the distribution of class labels in the original and balanced datasets. The plot compares the class distribution between the two datasets and saves it as a PDF if required.
}
\examples{
plot <- visualize_balance(original_data = original_data, balanced_data = balanced_data, group_col = "group")
plot <- visualize_balance(original_data = original_data, balanced_data = balanced_data, save_plots = FALSE)
}
