% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_08_normalize.R
\name{yeojohnson_transform}
\alias{yeojohnson_transform}
\title{Yeo-Johnson Transformation}
\usage{
yeojohnson_transform(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Transformed vector.
}
\description{
Yeo-Johnson Transformation
}
