% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_08_normalize.R
\name{z_score_standardize}
\alias{z_score_standardize}
\title{Z-Score Standardization}
\usage{
z_score_standardize(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Standardized vector.
}
\description{
Z-Score Standardization
}
