% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_08_clinical_analysis.R
\name{AttachClinicalData}
\alias{AttachClinicalData}
\title{Attach Clinical Data to Train_Model Object}
\usage{
AttachClinicalData(object, clinical_data)
}
\arguments{
\item{object}{A \code{Train_Model} object.}

\item{clinical_data}{A data frame with clinical variables. Rownames must
match the samples in the model's testing set.}
}
\value{
The updated \code{Train_Model} object.
}
\description{
Stores the entire clinical data frame inside \code{@process.info$clinical_data}.
Subsequent clinical analysis functions will use this by default.
}
