% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_evaluate_clustering.R
\name{calinski_harabasz}
\alias{calinski_harabasz}
\title{Calinski-Harabasz Index}
\usage{
calinski_harabasz(clustered.data)
}
\arguments{
\item{clustered.data}{Clustered data.}
}
\description{
Calinski-Harabasz Index
}
