% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_05_train.R
\name{check_factor_level}
\alias{check_factor_level}
\title{Check the Number of Levels in the Factor Variable}
\usage{
check_factor_level(data, group_col)
}
\arguments{
\item{data}{A data frame or tibble containing the dataset.}

\item{group_col}{The name of the column to check, which should represent class labels.}
}
\value{
TRUE if the column contains at least two levels, otherwise stops with an error message.
}
\description{
This function checks if the specified column in the provided dataset has at least two distinct levels. It is used to ensure that the validation data contains more than one class, which is a requirement for classification tasks.
}
\details{
This function ensures that the column representing class labels (\code{group_col}) has at least two levels, which is essential for classification problems. If the column has fewer than two levels, the function stops and throws an error indicating that the validation data is not suitable for classification.
}
\examples{
\dontrun{
# Check if the 'group' column in the dataset 'my_data' contains at least two levels
check_factor_level(data = my_data, group_col = "group")
}
}
