% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{convert_variables}
\alias{convert_variables}
\title{Convert Variables to Numeric or Factor Based on Their Types}
\usage{
convert_variables(
  data,
  variable_types,
  save_dir = here::here("StatObject", "Data"),
  save_data = F,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{data}{A data frame containing the variables to be converted.}

\item{variable_types}{A list containing the variable types. This should have a
component \code{numeric_vars}, which is a character vector of variable names that
should be converted to numeric.}
}
\value{
A data frame with the variables converted to the appropriate types
(numeric or factor).
}
\description{
This function converts the variables (columns) in the input data frame to either
numeric or factor types based on the information provided in \code{variable_types}.
Numeric variables are converted to numeric data type, and non-numeric variables
are converted to factors. The function uses \code{variable_types} to determine which
variables should be treated as numeric and which should be treated as factors.
}
\examples{
# Example of usage
data <- data.frame(a = c(1, 2, 3), b = c("low", "medium", "high"))
variable_types <- list(numeric_vars = c("a"))
converted_data <- convert_variables(data, variable_types)

}
