% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn1_create_object.R
\name{CreatePrognosiXObject}
\alias{CreatePrognosiXObject}
\title{Create PrognosiX Object}
\usage{
CreatePrognosiXObject(
  clean.data = NULL,
  info.data = data.frame(),
  time_col = "time",
  status_col = "status",
  baseline.table = NULL,
  variable.types = list(),
  survival.var = list(),
  survival.data = data.frame(),
  sub.data = data.frame(),
  univariate.analysis = list(),
  split.data = list(),
  feature.result = list(),
  filtered.set = list(),
  survival.model = NULL,
  best.model = list(),
  subgroup.risk = list(),
  object = NULL
)
}
\arguments{
\item{clean.data}{Clean data.}

\item{info.data}{Info data.}

\item{time_col}{Time col.}

\item{status_col}{Status col.}

\item{baseline.table}{Baseline table.}

\item{variable.types}{Variable types.}

\item{survival.var}{Survival var.}

\item{survival.data}{Survival data.}

\item{sub.data}{Sub data.}

\item{univariate.analysis}{Univariate analysis.}

\item{split.data}{Split data.}

\item{feature.result}{Feature result.}

\item{filtered.set}{Filtered set.}

\item{survival.model}{Survival model.}

\item{best.model}{Best model.}

\item{subgroup.risk}{Subgroup risk.}

\item{object}{Input object.}
}
\description{
Create PrognosiX Object
}
