% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_object.R
\name{CreateStatObject}
\alias{CreateStatObject}
\title{Create a Stat object for statistical analysis}
\usage{
CreateStatObject(
  raw.data = data.frame(),
  clean.data = data.frame(),
  info.data = data.frame(),
  group_col = "group",
  ...
)
}
\arguments{
\item{raw.data}{A data.frame containing the raw data for analysis. Defaults to an empty data frame.}

\item{clean.data}{A data.frame containing the cleaned data for analysis. Defaults to an empty data frame.}

\item{info.data}{A data.frame containing additional metadata related to the data. Defaults to an empty data frame.}

\item{group_col}{A character string specifying the name of the column used for grouping. Default is \code{"group"}.}

\item{...}{Additional arguments passed to methods (not used in this function).}
}
\value{
An object of class \code{Stat}, which contains the processed data and metadata as slots.
}
\description{
This function creates a \code{Stat} object, which is used to store and manage various components
of statistical analysis, such as raw data, cleaned data, additional metadata, and processing
information. It performs basic checks and preparation on the input data before creating the object
}
\examples{
# Creating a Stat object with example data
stat_obj <- CreateStatObject(raw.data = example_raw_data, clean.data = example_clean_data)

# Creating a Stat object with metadata
stat_obj <- CreateStatObject(raw.data = example_raw_data, clean.data = example_clean_data, info.data = example_info_data)
}
