% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub1_create_object.R
\name{CreateSubtypingObject}
\alias{CreateSubtypingObject}
\title{Create Subtyping Object}
\usage{
CreateSubtypingObject(
  clean.data = NULL,
  info.data = data.frame(),
  scale.data = data.frame(),
  Optimal.cluster = NULL,
  cluster.results = NULL,
  visualization.results = NULL,
  clustered.data = data.frame(),
  evaluation_results = list(),
  object = NULL
)
}
\arguments{
\item{clean.data}{Clean data.}

\item{info.data}{Info data.}

\item{scale.data}{Scale data.}

\item{Optimal.cluster}{Optimal cluster.}

\item{cluster.results}{Cluster results.}

\item{visualization.results}{Visualization results.}

\item{clustered.data}{Clustered data.}

\item{evaluation_results}{Evaluation results.}

\item{object}{Input object (Stat, Subtyping, PrognosiX, Model_data).}
}
\description{
Create Subtyping Object
}
