% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_evaluate_clustering.R
\name{davies_bouldin}
\alias{davies_bouldin}
\title{Davies-Bouldin Index}
\usage{
davies_bouldin(clustered.data)
}
\arguments{
\item{clustered.data}{Clustered data.}
}
\description{
Davies-Bouldin Index
}
