% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{detect_and_mark_outliers}
\alias{detect_and_mark_outliers}
\title{Detect and Mark Outliers in Data}
\usage{
detect_and_mark_outliers(
  data,
  group_col = "group",
  palette_name = "Royal1",
  save_plots = TRUE,
  save_dir = here("StatObject"),
  plot_display_num = 1,
  sub_var = NULL,
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)
}
\arguments{
\item{data}{A data frame containing the data to be processed. The data must include numeric variables.}

\item{group_col}{A string specifying the name of the grouping column (default is "group"). This column is
used to color the boxplots.}

\item{palette_name}{A string specifying the palette name for coloring the boxplots (default is "Royal1").}

\item{save_plots}{A logical value indicating whether to save the generated plots (default is TRUE).}

\item{save_dir}{A string specifying the directory to save the plots (default is "StatObject").}

\item{plot_display_num}{An integer specifying how many variables to display per plot (default is 5).}

\item{sub_var}{A vector of variable names to restrict the outlier detection and plotting to (default is NULL,
meaning all variables with outliers will be processed).}

\item{plot_width}{A numeric value specifying the width of the saved plot images (default is 5).}

\item{plot_height}{A numeric value specifying the height of the saved plot images (default is 5).}

\item{base_size}{A numeric value specifying the base size for the plot's text elements (default is 14).}
}
\value{
A list containing two elements:
\itemize{
\item \code{data_marked}: The original data frame with a new column for each detected outlier variable,
indicating whether each observation is an outlier.
\item \code{plots}: A list of ggplot objects representing the boxplots for visualizing outliers.
}
}
\description{
This function detects outliers in numeric variables of a given data frame. It identifies outliers using
the IQR (Interquartile Range) method and marks them in the data. The function generates boxplots to visualize
the detected outliers and saves these plots if specified.
}
\examples{
# Example 1: Detect and mark outliers, save plots to the default directory
result <- detect_and_mark_outliers(data = my_data)

# Example 2: Detect outliers and generate plots for specific variables only
result <- detect_and_mark_outliers(data = my_data, sub_var = c("var1", "var2"))
}
