% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_02_FeatureSelect.R
\name{ExplainModelPerformance}
\alias{ExplainModelPerformance}
\title{Explain Model Performance (ROC, Lift, Boxplot)}
\usage{
ExplainModelPerformance(
  explainer,
  geom = c("roc", "lift", "boxplot"),
  save_plots = FALSE,
  save_dir = "ModelExplain",
  plot_width = 6,
  plot_height = 5,
  ...
)
}
\arguments{
\item{explainer}{DALEX explainer object.}

\item{geom}{Plot type: "roc", "lift", or "boxplot".}

\item{save_plots}{Logical, save plot to file.}

\item{save_dir}{Directory to save the plot.}

\item{plot_width, plot_height}{Dimensions in inches.}

\item{...}{Additional arguments passed to `ggplot2::ggsave`.}
}
\value{
Invisibly, the DALEX model_performance object.
}
\description{
Explain Model Performance (ROC, Lift, Boxplot)
}
