% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_04_Extensions.R
\name{external_validation_test}
\alias{external_validation_test}
\title{Perform a survival version of the Hosmer Lemeshow test}
\usage{
external_validation_test(learner, object, time_point = NULL, n_groups = 10)
}
\arguments{
\item{learner}{A trained mlr3 learner (supports "distr").}

\item{object}{A `TaskSurv` or `PrognosiX` object (validation set).}

\item{time_point}{Time point for evaluation (default median event time).}

\item{n_groups}{Number of risk groups for HL test (default 10).}
}
\value{
A list with chi square statistic, p value, and observed/expected tables.
}
\description{
Perform a survival version of the Hosmer Lemeshow test
}
