% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-4-outlier.R
\name{extract_outlier_data}
\alias{extract_outlier_data}
\title{Extract Marked Outlier Data from Stat Object}
\usage{
extract_outlier_data(object, slot_name = "outlier.marked")
}
\arguments{
\item{object}{An object of class 'Stat' containing outlier information}

\item{slot_name}{Name of the slot containing outlier data (default: "outlier.marked")}
}
\value{
A data frame containing the original data with outlier markers (columns ending with "_outlier")
}
\description{
Retrieves the dataset with marked outliers from a specified slot in a 'Stat' class object.
This is a utility function for accessing pre-processed outlier information stored in S4 objects.
}
\examples{
\dontrun{
# Assuming 'object_stat' is a Stat object with outlier information
object_stat <- extract_outlier_data(object_stat)
}

}
