% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_07_outlier.R
\name{extract_outlier_data}
\alias{extract_outlier_data}
\title{Extract Outlier Data}
\usage{
extract_outlier_data(object)
}
\arguments{
\item{object}{Stat object.}
}
\description{
Extract Outlier Data
}
\examples{
\dontrun{
out_info <- extract_outlier_data (stat_obj_test)
}
}
