% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_4_outlier.R
\name{extract_outlier_data}
\alias{extract_outlier_data}
\title{Extract Outlier Data}
\usage{
extract_outlier_data(object)
}
\arguments{
\item{object}{Stat object.}
}
\description{
Extract Outlier Data
}
