% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_01_object.R, R/stat_08_normalize.R
\name{ExtractCleanData}
\alias{ExtractCleanData}
\title{Extract Clean Data from Stat Object}
\usage{
ExtractCleanData(object)

ExtractCleanData(object)
}
\arguments{
\item{object}{Stat object.}
}
\value{
A \code{data.frame} containing the cleaned data from the \code{Stat} object,
or \code{NULL} if extraction fails.
}
\description{
This function extracts the \code{clean.data} slot from an object of class \code{Stat}.
}
\examples{
\dontrun{
data <- ExtractCleanData(stat_obj_test)
}
}
