% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_01_object.R, R/stat_08_normalize.R
\name{ExtractCleanData}
\alias{ExtractCleanData}
\title{Extract Clean Data from Stat Object}
\usage{
ExtractCleanData(object)

ExtractCleanData(object)
}
\arguments{
\item{object}{Stat object.}
}
\value{
A `data.frame` containing the cleaned data from the `Stat` object,
or `NULL` if extraction fails.
}
\description{
This function extracts the `clean.data` slot from an object of class `Stat`.
}
\examples{
\dontrun{
data <- ExtractCleanData(stat_obj_test)
}
}
