% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_5_normalize.R,
%   R/module_2_model_object.R
\name{ExtractCleanData}
\alias{ExtractCleanData}
\title{Extract Clean Data}
\usage{
ExtractCleanData(object)

ExtractCleanData(object)
}
\arguments{
\item{object}{An object of class \code{Stat}.}
}
\value{
A \code{data.frame} containing the cleaned data from the \code{Stat} object,
or \code{NULL} if extraction fails.
}
\description{
This function extracts the \code{clean.data} slot from an object of class \code{Stat}.
}
\examples{
clean_df <- ExtractCleanData(my_stat_object)

}
