% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub1_create_object.R
\name{ExtractInfoData}
\alias{ExtractInfoData}
\title{Extract Info Data}
\usage{
ExtractInfoData(object)
}
\arguments{
\item{object}{An object containing info.data slot.}
}
\description{
Extract Info Data
}
