% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_02_FeatureSelect.R
\name{FeatureSelectSBF}
\alias{FeatureSelectSBF}
\title{Univariate Filter Feature Selection for Train_Model objects}
\usage{
FeatureSelectSBF(
  object,
  sbf_func = caret::rfSBF,
  method = "repeatedcv",
  number = 5,
  repeats = 2,
  metric = "Accuracy",
  allowParallel = FALSE,
  save_plot = FALSE,
  save_dir = NULL,
  seed = 825
)
}
\arguments{
\item{object}{A Train_Model or Stat object.}

\item{sbf_func}{A caret SBF function list (e.g., \code{rfSBF}, \code{caretSBF}).}

\item{method}{External resampling method.}

\item{number}{Number of folds.}

\item{repeats}{Repeats.}

\item{metric}{Evaluation metric.}

\item{allowParallel}{Logical.}

\item{save_plot}{Logical (saves variable count barplot).}

\item{save_dir}{Directory.}

\item{seed}{Random seed.}
}
\value{
List with \code{result} (sbf object), \code{opt_vars}, \code{plot}.
}
\description{
Univariate Filter Feature Selection for Train_Model objects
}
