% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_07_outlier.R
\name{handle_outliers}
\alias{handle_outliers}
\title{Handle Outliers}
\usage{
handle_outliers(
  data,
  outlier_info,
  method = "remove",
  impute_value = "median",
  save_dir = NULL,
  save_data = FALSE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{data}{A data frame.}

\item{outlier_info}{Outlier information list.}

\item{method}{Handling method ("remove", "winsorize", "impute").}

\item{impute_value}{Value to impute (e.g., median).}

\item{save_dir}{Directory to save cleaned data.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Handle Outliers
}
\examples{
\dontrun{
out_info <- extract_outlier_data (stat_obj_test)
clean_after <- handle_outliers(stat_obj_test@clean.data, out_info, method = "remove", save_data = FALSE)
}
}
