% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_4_outlier.R
\name{handle_outliers}
\alias{handle_outliers}
\title{Handle Outliers}
\usage{
handle_outliers(
  data,
  outlier_info,
  method = "remove",
  impute_value = "median",
  save_dir = here::here("StatObject", "Data"),
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{data}{A data frame.}

\item{outlier_info}{Outlier information list.}

\item{method}{Handling method ("remove", "winsorize", "impute").}

\item{impute_value}{Value to impute (e.g., median).}

\item{save_dir}{Directory to save cleaned data.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Handle Outliers
}
