% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute Missing Values in a Data Frame}
\usage{
impute_missing_values(
  data,
  group_col = "group",
  impute_method = "mice",
  m = 5,
  max_unique_values = 5
)
}
\arguments{
\item{data}{A data frame containing the dataset to be imputed.}

\item{group_col}{A string specifying the name of the grouping column (default is "group"). This column is
excluded from the imputation process.}

\item{impute_method}{A string specifying the imputation method to be used. Options are:
"mice" (multiple imputation using MICE) or "median_mode" (simple median for numeric
and mode for categorical variables). Default is "mice".}

\item{m}{An integer specifying the number of imputed datasets when using the MICE method (default is 5).}

\item{max_unique_values}{An integer specifying the maximum number of unique values a variable can have
to be considered categorical. Default is 5.}
}
\value{
A data frame with missing values imputed according to the selected method.
}
\description{
This function imputes missing values in a data frame using either multiple imputation via MICE or
simple median and mode imputation. The method to be used is specified by the \code{impute_method} parameter.
Numeric and categorical variables are handled accordingly, with numeric values imputed using the median
and categorical values imputed using the mode.
}
\examples{
# Example 1: Impute missing values using the MICE method with 5 imputed datasets
imputed_data <- impute_missing_values(my_data, impute_method = "mice", m = 5)

# Example 2: Impute missing values using the median and mode imputation method
imputed_data <- impute_missing_values(my_data, impute_method = "median_mode")
}
