% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub3_kmeans.R
\name{kmeans_clustering}
\alias{kmeans_clustering}
\title{K-means Clustering}
\usage{
kmeans_clustering(data, k, nstart = 50)
}
\arguments{
\item{data}{Numeric data.}

\item{k}{Number of clusters.}

\item{nstart}{nstart.}
}
\description{
K-means Clustering
}
