% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub3_kmeans.R
\name{kmeans_with_optimal_k}
\alias{kmeans_with_optimal_k}
\title{K-means with Optimal K}
\usage{
kmeans_with_optimal_k(
  data,
  palette_name = "Zissou1",
  save_plots = TRUE,
  save_dir = here("Subtyping", "kmeans_result"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  nstart = 50,
  seed = 123
)
}
\arguments{
\item{data}{Data frame.}

\item{palette_name}{Palette.}

\item{save_plots}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Base size.}

\item{nstart}{nstart.}

\item{seed}{Seed.}
}
\description{
K-means with Optimal K
}
