% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{log_transform}
\alias{log_transform}
\title{Log Transformation of Input Data}
\usage{
log_transform(x)
}
\arguments{
\item{x}{A numeric vector to be log-transformed. The function assumes that the
input contains non-negative values.}
}
\value{
A list containing:
\itemize{
\item \code{scaled_data}: A numeric vector of log-transformed values.
\item \code{normalize_method}: A string indicating the normalization method used (log transformation).
\item \code{normalize_info}: An empty list that can be used to store additional normalization details.
}
}
\description{
This function applies a log transformation to the input vector \code{x}. The function
adds a small constant (0.00000001) to each value before applying the logarithm to
avoid taking the log of zero. It checks that the input does not contain negative
values, as log transformation is not valid for negative values.
}
\examples{
# Example of using log_transform
x <- c(1, 2, 3, 10)
result <- log_transform(x)
result$scaled_data  # The log-transformed values

}
