% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_5_normalize.R
\name{log_transform}
\alias{log_transform}
\title{Log Transformation}
\usage{
log_transform(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Log-transformed vector.
}
\description{
Log Transformation
}
