% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_lpa.R
\name{lpa_with_optimal_k}
\alias{lpa_with_optimal_k}
\title{LPA with Optimal K}
\usage{
lpa_with_optimal_k(
  data,
  max_clusters = 5,
  model_names = c("EII", "VII", "EEI", "VEI", "EVI", "VVI"),
  save_plots = TRUE,
  save_dir = here("Subtyping", "lpa_result"),
  plot_width = 7,
  plot_height = 5,
  base_size = 14,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data.}

\item{max_clusters}{Max clusters.}

\item{model_names}{Model names.}

\item{save_plots}{Save plots.}

\item{save_dir}{Save dir.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Base size.}

\item{seed}{Seed.}

\item{verbose}{Verbose.}
}
\description{
LPA with Optimal K
}
