% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-17-cm.R
\name{match_factor_levels}
\alias{match_factor_levels}
\title{Match Factor Levels Between External and Reference Datasets}
\usage{
match_factor_levels(external_data, reference_data)
}
\arguments{
\item{external_data}{A data frame containing new data (e.g., test or validation set)
to be aligned in terms of factor levels.}

\item{reference_data}{A data frame used as the reference for factor levels
(usually the training dataset).}
}
\value{
A modified version of \code{external_data} where factor variables
that also appear in \code{reference_data} are converted to factors
with matched levels.
}
\description{
This function ensures that factor variables in an external dataset have the same levels
as those in a reference dataset (e.g., training set). This is particularly useful
before applying a trained model to new data to avoid factor level mismatches.
}
\examples{
# Assume train_data and test_data share some columns
test_data <- match_factor_levels(test_data, train_data)

}
