% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-21-rocall.R
\name{ModelBestRoc}
\alias{ModelBestRoc}
\title{Generate and Save ROC Plots for the Best Model in a 'Best_Model' Object}
\usage{
ModelBestRoc(
  object,
  group_col = "group",
  palette_name = "AsteroidCity1",
  base_size = 14,
  save_plots = TRUE,
  save_dir = here("ModelData", "best_model_result"),
  plot_width = 5,
  plot_height = 5,
  alpha = 1
)
}
\arguments{
\item{object}{An object of class 'Best_Model' that contains the best model and associated datasets.}

\item{group_col}{The name of the grouping/label column. Default is "group". If the object is of class 'Best_Model', this will be overwritten.}

\item{palette_name}{The name of the color palette to use from the wesanderson package. Default is "AsteroidCity1".}

\item{base_size}{The base font size for plots. Default is 14.}

\item{save_plots}{Logical; whether to save the ROC plot as a PDF. Default is TRUE.}

\item{save_dir}{Directory where the ROC plot will be saved. Default is \code{here("ModelData", "best_model_result")}.}

\item{plot_width}{Width of the saved plot in inches. Default is 5.}

\item{plot_height}{Height of the saved plot in inches. Default is 5.}

\item{alpha}{Transparency level for plot elements. Default is 1 (opaque).}
}
\value{
An updated 'Best_Model' object with ROC plot data stored in the \code{best.model.result[["roc_plots"]]} slot.
}
\description{
This function extracts the best model and relevant datasets from a 'Best_Model' object,
computes ROC curves across different validation cohorts (training, testing, validation, external validation),
and visualizes the ROC curves. It also updates the 'Best_Model' object with the resulting ROC plot data.
}
\examples{
\dontrun{
  object_best <- ModelBestRoc(object = object_best)
}
}
