% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Data with Various Methods}
\usage{
normalize_data(
  data,
  normalize_method = "min_max_scale",
  group_col = "group",
  max_unique_values = 5
)
}
\arguments{
\item{data}{A data frame containing the data to be normalized. Numeric columns will be transformed
based on the selected normalization method.}

\item{normalize_method}{A string specifying the normalization method. One of: "log_transform",
"min_max_scale", "z_score_standardize", "max_abs_scale", "center_data",
or "scale_data". Default is "min_max_scale".}

\item{group_col}{A string specifying the column that contains group information. This is used for
diagnostic purposes and is optional. Default is "group".}

\item{max_unique_values}{The maximum number of unique values to classify a column as categorical.
Default is 5. Used in \code{diagnose_variable_type} to classify variables.}
}
\value{
A list containing:
\itemize{
\item \code{scaled_data}: A data frame with normalized values for the numeric columns.
\item \code{normalize_method}: The method used for normalization (as a string).
\item \code{normalize_info}: A list containing additional information related to the normalization process.
}
}
\description{
This function normalizes the numeric columns in the provided data frame using a selected method.
The available methods include log transformation, min-max scaling, Z-score standardization,
centering, and scaling. It also ensures that non-numeric columns (e.g., categorical) are preserved
during normalization.
}
\examples{
# Example of using normalize_data
data <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6), group = c("X", "Y", "X"))
result <- normalize_data(data, normalize_method = "z_score_standardize")
result$scaled_data  # The normalized data
result$normalize_method  # Method used for normalization

}
