% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_08_normalize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Data}
\usage{
normalize_data(
  data,
  method = "auto",
  group_col = "group",
  max_unique_values = 5,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "scale_data.csv"
)
}
\arguments{
\item{data}{Data frame.}

\item{method}{Normalization method ("auto" or specific).}

\item{group_col}{Group column.}

\item{max_unique_values}{Max unique values.}

\item{save_dir}{Save directory.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\value{
Normalized data frame.
}
\description{
Normalize Data
}
\examples{
\dontrun{
nor <- normalize_data(stat_obj_test@clean.data, method = "z_score", group_col = "SWAB")
print(nor)
}
}
