% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-Hot Encoding for Categorical Variables}
\usage{
one_hot_encode(data, group_col = "group", max_unique_values = 5)
}
\arguments{
\item{data}{A data frame containing the data to be encoded.}

\item{group_col}{A string specifying the column that contains group information. This is optional and,
if provided, will be moved to the end of the encoded data frame. Default is "group".}

\item{max_unique_values}{The maximum number of unique values a variable can have to be considered categorical.
Default is 5. This is used in \code{diagnose_variable_type} to determine which columns
should be one-hot encoded.}
}
\value{
A data frame where categorical variables are replaced by one-hot encoded columns.
The group column, if provided, is placed at the end of the data frame.
}
\description{
This function performs one-hot encoding on categorical variables in the provided data frame.
It identifies categorical variables based on the number of unique values and encodes them
into binary columns where each unique value of the categorical variable becomes a separate column.
If a group column is provided, it is moved to the end of the data frame.
}
\examples{
# Example of one-hot encoding on a data frame with a categorical variable
data <- data.frame(
  group = c("A", "B", "A", "B", "C"),
  category = c("X", "Y", "X", "Z", "Y")
)
encoded_data <- one_hot_encode(data, group_col = "group")
encoded_data

}
