% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-Hot Encode Categorical Variables in a Data Frame}
\usage{
one_hot_encode(
  data,
  group_col = "group",
  max_unique_values = 5,
  save_dir = here::here("StatObject", "Data"),
  save_data = F,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{data}{A data frame to be processed.}

\item{group_col}{The column name for grouping (e.g., "group"). This column will not be encoded
and will be moved to the end of the result. Default is \code{"group"}.}

\item{max_unique_values}{The maximum number of unique values a variable can have to be
considered for one-hot encoding. Default is 5.}
}
\value{
A new data frame with one-hot encoded variables.
}
\description{
This function performs one-hot encoding on categorical variables with limited unique values,
as determined by the \code{diagnose_variable_type} function. It skips encoding the \code{group_col}
and moves it to the last column of the resulting data frame if provided.
}
\examples{
\dontrun{
one_hot_encode(iris, group_col = "Species", max_unique_values = 5)
}
}
