% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_04_descriptive.R, R/utils_convert_types.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-Hot Encode Categorical Variables in a Data Frame}
\usage{
one_hot_encode(
  data,
  group_col = "group",
  max_unique_values = 5,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)

one_hot_encode(
  data,
  group_col = "group",
  max_unique_values = 5,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{data}{A data frame or an S4 object (Stat, Train_Model, Subtyping, PrognosiX).}

\item{group_col}{The name of the group column.}

\item{max_unique_values}{Maximum unique values for encoding.}

\item{save_dir}{Directory to save.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\value{
A new data frame with one-hot encoded variables.
}
\description{
This function performs one-hot encoding on categorical variables with limited unique values,
as determined by the `diagnose_variable_type` function. It skips encoding the `group_col`
and moves it to the last column of the resulting data frame if provided.
}
\examples{
\dontrun{
one_hot_encode(iris, group_col = "Species", max_unique_values = 2,save_dir = ".")
}
\dontrun{
encoded <- one_hot_encode(stat_obj_test@clean.data, group_col = "SWAB", save_data = FALSE)
}
}
