% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub5_tsne_vis.R
\name{perform_tsne}
\alias{perform_tsne}
\title{Perform t-SNE}
\usage{
perform_tsne(data, perplexity = 30, dims = 2, iterations = 1000)
}
\arguments{
\item{data}{Data.}

\item{perplexity}{Perplexity.}

\item{dims}{Dims.}

\item{iterations}{Iterations.}
}
\description{
Perform t-SNE
}
