% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_3_descriptive.R
\name{plot_categorical_descriptive}
\alias{plot_categorical_descriptive}
\title{Plot Categorical Descriptive Statistics for 'Stat' Object or Data Frame}
\usage{
plot_categorical_descriptive(
  object,
  group_col = "group",
  palette_name = "Royal1",
  save_plots = TRUE,
  save_dir = here("StatObject", "categorical_descriptive"),
  sub_var = NULL,
  plot_width = 5,
  plot_height = 5,
  base_size = 14
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame containing the dataset to analyze.}

\item{group_col}{The column name representing the grouping variable (default is "group").}

\item{palette_name}{The name of the color palette to use (default is "Royal1").}

\item{save_plots}{A logical value indicating whether to save the plots (default is TRUE).}

\item{save_dir}{The directory where the plots will be saved (default is "StatObject/categorical_descriptive").}

\item{sub_var}{A vector of column names to subset the variables to plot (default is NULL, meaning all categorical variables will be plotted).}

\item{plot_width}{The width of the saved plot (default is 5).}

\item{plot_height}{The height of the saved plot (default is 5).}

\item{base_size}{The base font size for the plots (default is 14).}
}
\value{
If the input is a 'Stat' object, the function returns the updated 'Stat' object with a new slot \code{compute.descriptive$count_plots}
containing the generated plots. If the input is a data frame, it returns a list of the generated plots.
}
\description{
This function generates bar plots displaying the distribution of categorical variables and their percentages across different groups.
It can handle both \code{Stat} objects and standard data frames. The function also supports saving plots to a specified directory.
The plots are created using \code{ggplot2} and the colors are customized using a specified color palette.
}
\examples{
# Example 1: Plot categorical descriptive statistics for a 'Stat' object
updated_stat <- plot_categorical_descriptive(object = my_stat, group_col = "group")

# Example 2: Plot categorical descriptive statistics for a data frame
plot_results <- plot_categorical_descriptive(object = my_data, group_col = "group")
}
