% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_06_deg.R
\name{plot_deg_boxplot}
\alias{plot_deg_boxplot}
\title{Generate box Plot for Differentially Expressed Genes}
\usage{
plot_deg_boxplot(
  last_test_sig,
  data,
  control = "health",
  case = "cancer",
  top_n = 5,
  palette_name = "Royal1",
  name_identity = "deg",
  save_plots = TRUE,
  save_dir = NULL,
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  title = "Violin Plot",
  group_col = "group",
  pseudocount = 1
)
}
\arguments{
\item{last_test_sig}{A data frame containing the results of differential expression analysis with columns
such as 'change' (e.g., 'Stable', 'Upregulated', 'Downregulated') and 'logFC' (log-fold change).}

\item{data}{A data frame containing the expression data with rows as samples and columns as features.}

\item{control}{The label of the control group (default: "health").}

\item{case}{The label of the case group (default: "cancer").}

\item{top_n}{The number of top significant features to display (default: 5).}

\item{palette_name}{The name of the color palette for the plot.}

\item{name_identity}{The identity for the type of analysis, default is "deg".}

\item{save_plots}{Logical value to indicate whether to save the plot (default: TRUE).}

\item{save_dir}{The directory to save the plot (default: "here('StatObject', 'deg_info')").}

\item{plot_width}{The width of the plot in inches (default: 5).}

\item{plot_height}{The height of the plot in inches (default: 5).}

\item{base_size}{The base font size for the plot (default: 14).}

\item{title}{The title of the plot (default: "Violin Plot").}

\item{group_col}{The column name used to group the data (default: 'group').}

\item{pseudocount}{A small positive value added before log10 transform to avoid
log(0) / NaN. Default: 1 (i.e. log10(value + 1)).}
}
\value{
A ggplot object representing the violin plot.
}
\description{
This function generates a violin plot for the top differentially expressed genes based on their log-fold change
and statistical significance. It also calculates the Wilcoxon test p-values for group comparisons and annotates
the plot with significance stars. The plot can be saved to a specified directory.
}
\examples{
\dontrun{
stat_obj <- stat_var_feature(stat_obj_test)
last_sig <- ExtractLastTestSig(stat_obj)
print(last_sig)
plot_deg_boxplot(last_test_sig = last_sig, data = stat_obj@clean.data,group='SWAB',save_dir = "./")
plot_deg_boxplot(last_test_sig = last_sig,data = stat_obj@clean.data,group='SWAB',save_dir = "./",top_n = 3)
}
}
