% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_06_deg.R
\name{plot_deg_radarchart}
\alias{plot_deg_radarchart}
\title{Plot DEG Radar Chart}
\usage{
plot_deg_radarchart(
  df,
  palette_name = "Zissou1",
  x_col = "id",
  y_cols = c("mean_x", "mean_y"),
  size_col = "logFC",
  color_col = "logp",
  fill_col = "change",
  p_adjust_col = "p.adjust",
  plot_width = 5,
  plot_height = 5,
  save_dir = NULL,
  base_size = 14,
  title = "Radar Chart Title"
)
}
\arguments{
\item{df}{A data frame containing the DEG results with columns for ID, means, log-fold change, p-values,
adjusted p-values, and other necessary data.}

\item{palette_name}{A string indicating the name of the color palette to use for the chart .}

\item{x_col}{The column name to use as the x-axis (default: "id").}

\item{y_cols}{A vector of column names to be plotted on the y-axis (default: c("mean_x", "mean_y")).}

\item{size_col}{The column name for the size of the points (default: "logFC").}

\item{color_col}{The column name for the color of the points, typically for p-values (default: "logp").}

\item{fill_col}{The column name for the fill color, which defines the grouping (default: "change").}

\item{p_adjust_col}{The column name for the adjusted p-values (default: "p.adjust").}

\item{plot_width}{The width of the plot (default: 5).}

\item{plot_height}{The height of the plot (default: 5).}

\item{save_dir}{The directory where the radar chart image will be saved (default: "here('StatObject', 'deg_info')").}

\item{base_size}{The base font size for the plot (default: 14).}

\item{title}{The title of the radar chart (default: "Radar Chart Title").}
}
\value{
The ggplot object representing the radar chart.
}
\description{
This function generates a radar chart visualizing the differential expression (DEG) results from a data frame.
It uses a set of custom parameters to customize the appearance, such as colors, size, and grouping. The chart can be
saved to a specified directory as a PDF file.
}
\examples{
\dontrun{
stat_obj <- stat_var_feature(stat_obj_test)
last_sig <- ExtractLastTestSig(stat_obj)
print(last_sig)
plot_deg_radarchart(last_sig, palette_name = "Zissou1", title = "Differential Expression Radar")
}
}
