% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_03_feature_sensitivity.R
\name{plot_elbow}
\alias{plot_elbow}
\title{Plot elimination curve with optional optimal points}
\usage{
plot_elbow(
  elim_obj,
  best_features = NULL,
  facet = FALSE,
  palette_name = "Darjeeling1",
  use_low_sat = TRUE,
  ci_style = c("errorbar", "ribbon", "none"),
  save_plot = FALSE,
  save_dir = NULL,
  width = if (facet) 12 else 8,
  height = 5,
  ncol = NULL,
  ...
)
}
\description{
Plot elimination curve with optional optimal points
}
