% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-8.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot ROC Curves and Compute AUC Values}
\usage{
plot_roc_curve(
  model_list,
  validation_data,
  group_col = "group",
  palette_name = "AsteroidCity1",
  base_size = 14,
  save_plots = TRUE,
  save_dir = here("ModelData", "best.model.result"),
  plot_width = 5,
  plot_height = 5,
  alpha = 1
)
}
\arguments{
\item{model_list}{A list containing multiple models. Each element in the list is a trained model.}

\item{validation_data}{A dataset used to validate the model performance. The dataset should include both actual labels and features.}

\item{group_col}{The column name in the validation data representing the actual labels, default is "group".}

\item{palette_name}{The color palette name used for the plot, default is "AsteroidCity1".}

\item{base_size}{The base font size for the plot, default is 14.}

\item{save_plots}{Whether to save the generated plots, default is TRUE.}

\item{save_dir}{The directory path to save the plot, default is \code{here("ModelData", "best.model.result")}.}

\item{plot_width}{The width of the plot in inches, default is 5.}

\item{plot_height}{The height of the plot in inches, default is 5.}

\item{alpha}{The transparency level of the curve, default is 1 (fully opaque).}
}
\value{
A list containing the ROC objects for each model.
}
\description{
This function plots the ROC curves for multiple models and computes the AUC (Area Under the Curve) value along with its confidence interval for each model. The ROC curve is used to evaluate the performance of classification models, where a higher AUC value indicates a better model.
}
\details{
This function evaluates the performance of each model by calculating the ROC curve and AUC value. It generates a plot with all the ROC curves of the models and labels each curve with the AUC value and its 95\% confidence interval in the legend.
The user can choose to save the plot as a PDF file, and the function returns a list containing the ROC objects for further analysis.
}
\examples{
# Call this function with a list of trained models and validation data
plot_roc_curve(model_list = my_model_list,
               validation_data = my_validation_data,
               group_col = "group",
               palette_name = "AsteroidCity1",
               save_plots = TRUE,
               save_dir = "path/to/save",
               plot_width = 7,
               plot_height = 7)

}
