% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_06_ensemble.R
\name{PlotProbDensity}
\alias{PlotProbDensity}
\title{Prediction Probability Density Plot (fixed)}
\usage{
PlotProbDensity(model_obj, data = NULL, save_plot = FALSE, save_dir = NULL)
}
\arguments{
\item{model_obj}{A trained Train_Model object.}

\item{data}{Test data (default: split.data$testing).}

\item{save_plot}{Logical.}

\item{save_dir}{Output directory.}
}
\description{
Prediction Probability Density Plot (fixed)
}
