% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_08_clinical_analysis.R
\name{PlotSubgroupForest}
\alias{PlotSubgroupForest}
\title{Subgroup Performance Forest Plot}
\usage{
PlotSubgroupForest(
  model_obj,
  clinical_data = NULL,
  subgroup_var,
  newdata = NULL,
  model_name = NULL,
  save_plot = FALSE,
  save_dir = NULL,
  palette_name = "Darjeeling1"
)
}
\arguments{
\item{model_obj, clinical_data, subgroup_var, newdata, model_name}{As above.}

\item{save_plot, save_dir, palette_name}{Output options.}
}
\value{
A ggplot object.
}
\description{
Computes AUC with 95\\% CI within each subgroup defined by a categorical
clinical variable.
}
