% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_08_clinical_analysis.R
\name{PlotThresholdDensity}
\alias{PlotThresholdDensity}
\title{Decision Density with Threshold Zones}
\usage{
PlotThresholdDensity(
  thresh_result,
  lower_threshold = NULL,
  upper_threshold = NULL,
  save_plot = FALSE,
  save_dir = NULL
)
}
\arguments{
\item{thresh_result}{Output from \code{CalculateThresholds}.}

\item{lower_threshold}{Default: \code{Se_Target} if exists, else Youden.}

\item{upper_threshold}{Default: \code{Sp_Target} if exists, else Youden.}

\item{save_plot, save_dir}{Output options.}
}
\value{
A ggplot object.
}
\description{
Decision Density with Threshold Zones
}
