% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4_pn1_create_object.R
\docType{class}
\name{PrognosiX-class}
\alias{PrognosiX-class}
\alias{PrognosiX}
\title{PrognosiX Class}
\description{
PrognosiX Class
}
\section{Slots}{

\describe{
\item{\code{clean.data}}{Data frame.}

\item{\code{info.data}}{Data frame.}

\item{\code{time_col}}{Any.}

\item{\code{status_col}}{Any.}

\item{\code{survival.data}}{Data frame.}

\item{\code{baseline.table}}{Any.}

\item{\code{variable.types}}{List.}

\item{\code{survival.var}}{List.}

\item{\code{sub.data}}{Data frame.}

\item{\code{univariate.analysis}}{List.}

\item{\code{split.data}}{List.}

\item{\code{feature.result}}{List.}

\item{\code{filtered.set}}{List.}

\item{\code{survival.model}}{Any.}

\item{\code{best.model}}{List.}

\item{\code{subgroup.risk}}{List.}
}}

