% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{remove_high_missing}
\alias{remove_high_missing}
\title{Remove Variables and Samples with Excessive Missing Data}
\usage{
remove_high_missing(data, miss_threshold = 20)
}
\arguments{
\item{data}{A data frame containing the dataset to be processed.}

\item{miss_threshold}{A numeric value between 0 and 100 specifying the threshold (in percentage) for
missing data. Variables or samples with missing data above this threshold will
be removed (default is 20).}
}
\value{
A data frame with variables and samples removed if their missing data exceeds the given threshold.
}
\description{
This function removes variables and samples from the input data frame where the percentage of missing
values exceeds the specified threshold. Variables with missing data above the threshold will be removed,
and the same applies to samples (rows) where the percentage of missing data exceeds the threshold.
}
\examples{
# Example 1: Remove variables and samples with more than 20\% missing data
cleaned_data <- remove_high_missing(my_data, miss_threshold = 20)

# Example 2: Remove variables and samples with more than 10\% missing data
cleaned_data_10 <- remove_high_missing(my_data, miss_threshold = 10)
}
