% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_03_feature_sensitivity.R
\name{run_feature_elimination}
\alias{run_feature_elimination}
\title{Run Backward Feature Elimination}
\usage{
run_feature_elimination(
  object,
  models = c("rf", "gbm", "glmboost"),
  metric = "ROC",
  number = 5,
  seed = 825,
  min_features = 2,
  smooth_span = 0,
  verbose = TRUE,
  parallel = FALSE,
  n_cores = NULL
)
}
\description{
Run Backward Feature Elimination
}
