% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_05_train.R
\name{SelectBestModel}
\alias{SelectBestModel}
\title{Select Best Performing Model (Fixed)}
\usage{
SelectBestModel(object, metric = "auc", custom_selection = NULL)
}
\arguments{
\item{object}{A Train_Model object.}

\item{metric}{Primary metric (default "auc").}

\item{custom_selection}{Optional model name to force selection.}
}
\value{
Updated Train_Model object.
}
\description{
Select Best Performing Model (Fixed)
}
