% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_evaluate_clustering.R
\name{silhouette_score}
\alias{silhouette_score}
\title{Silhouette Score}
\usage{
silhouette_score(clustered.data)
}
\arguments{
\item{clustered.data}{Clustered data.}
}
\description{
Silhouette Score
}
