% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_3_descriptive.R
\name{stat_compute_descriptive}
\alias{stat_compute_descriptive}
\title{Compute Descriptive Statistics for 'Stat' Object or Data Frame}
\usage{
stat_compute_descriptive(
  object,
  count_feature = TRUE,
  group_col = "group",
  max_unique_values = 5
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame containing the dataset to analyze.}

\item{count_feature}{A logical value indicating whether to compute counts for categorical variables (default is TRUE).}

\item{group_col}{The column name representing the grouping variable (default is "group").}

\item{max_unique_values}{The maximum number of unique values a variable can have to be considered categorical (default is 5).}
}
\value{
If the input is a 'Stat' object, the function returns the updated 'Stat' object with a new slot \code{compute.descriptive}
containing the results. If the input is a data frame, it returns the descriptive statistics as a list.
}
\description{
This function computes descriptive statistics for numeric and categorical variables in the input data. It can handle both
\code{Stat} objects and standard data frames. The function computes counts for categorical variables, and summary statistics
(mean, median, standard deviation, min, max) for numeric variables, along with normality tests for numeric variables.
}
\examples{
# Example 1: Compute descriptive statistics for a 'Stat' object
updated_stat <- stat_compute_descriptive(object = my_stat, count_feature = TRUE, group_col = "group")

# Example 2: Compute descriptive statistics for a data frame
result <- stat_compute_descriptive(object = my_data, count_feature = TRUE)
}
