% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_07_outlier.R
\name{stat_detect_and_mark_outliers}
\alias{stat_detect_and_mark_outliers}
\title{Detect and Mark Outliers in Stat Object}
\usage{
stat_detect_and_mark_outliers(
  object,
  method = "zscore",
  threshold = 3,
  group_col = "group",
  max_unique_values = 5,
  save_dir = NULL,
  save_data = FALSE,
  csv_filename = "outlier_info.csv"
)
}
\arguments{
\item{object}{Stat object.}

\item{method}{Detection method.}

\item{threshold}{Threshold.}

\item{group_col}{Group column.}

\item{max_unique_values}{Max unique values.}

\item{save_dir}{Directory to save results.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Detect and Mark Outliers in Stat Object
}
\examples{
\dontrun{
out_info <- stat_detect_and_mark_outliers(stat_obj_test, method = "zscore", threshold = 3)
}
}
