% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_2_missing_plot.R
\name{stat_gaze_analysis}
\alias{stat_gaze_analysis}
\title{Statistical Gaze Analysis}
\usage{
stat_gaze_analysis(
  object,
  formula = NULL,
  group_col = "group",
  digits = 1,
  show.p = TRUE,
  gaze_method = 3,
  save_word = TRUE,
  save_dir = here("StatObject", "gaze_baseline")
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame containing the data to analyze.}

\item{formula}{A formula specifying the model to fit. If NULL, a default formula is used (default is NULL).}

\item{group_col}{The column name representing the grouping variable (default is "group").}

\item{digits}{The number of digits to display for the result (default is 1).}

\item{show.p}{A logical value indicating whether to display p-values in the output (default is TRUE).}

\item{gaze_method}{An integer between 1 and 5 representing the gaze analysis method to use (default is 3).}

\item{save_word}{A logical value indicating whether to save the results to a Word document (default is TRUE).}

\item{save_dir}{The directory to save the Word document and plot images (default is the "StatObject/gaze_baseline" folder).}
}
\value{
An updated 'Stat' object with the gaze analysis results if the input is of class 'Stat',
or a data frame containing the results otherwise.
}
\description{
This function performs gaze analysis on a given dataset or \code{Stat} object, providing results for
group comparisons, statistical significance, and optionally saving the results as a Word document
and plots as images. It supports customized formulas and gaze methods for the analysis.
}
\examples{
stat_object <- Stat$new(clean.data = your_data_frame, group_col = "group")
updated_stat_object <- stat_gaze_analysis(stat_object,
                                         formula = ~ group + age + gender,
                                         digits = 2,
                                         show.p = TRUE,
                                         gaze_method = 3,
                                         save_word = TRUE,
                                         save_plots = TRUE,
                                         save_dir = "path/to/save")

# Example 2: Performing gaze analysis on a data frame
result <- stat_gaze_analysis(your_data_frame,
                             formula = ~ group + age + gender,
                             save_word = FALSE,
                             save_plots = TRUE)

}
