% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_07_outlier.R
\name{stat_handle_outliers}
\alias{stat_handle_outliers}
\title{Handle Outliers in Stat Object}
\usage{
stat_handle_outliers(
  object,
  method = "remove",
  impute_value = "median",
  save_dir = NULL,
  save_data = FALSE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{object}{Stat object.}

\item{method}{Handling method.}

\item{impute_value}{Impute value.}

\item{save_dir}{Save directory.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Handle Outliers in Stat Object
}
\examples{
\dontrun{
stat_obj <- stat_handle_outliers(stat_obj_test, method = "remove")
stat_obj <- stat_handle_outliers(stat_obj_test, method = "impute")
}
}
