% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_03_remove_missing.R
\name{stat_miss_remove}
\alias{stat_miss_remove}
\title{Remove Missing Values from Stat Object or Data Frame}
\usage{
stat_miss_remove(
  object,
  miss_threshold = 25,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{object}{A Stat object or data frame.}

\item{miss_threshold}{Missing threshold.}

\item{save_dir}{Directory to save.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Remove Missing Values from Stat Object or Data Frame
}
\examples{
\dontrun{
stat_obj<- stat_miss_remove(stat_obj_test, miss_threshold = 25)
}
}
