% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_2_remove_miss.R
\name{stat_miss_remove}
\alias{stat_miss_remove}
\title{Remove Missing Values from Stat Object or Data Frame}
\usage{
stat_miss_remove(
  object,
  miss_threshold = 25,
  save_dir = here::here("StatObject", "Data"),
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{object}{A Stat object or data frame.}

\item{miss_threshold}{Missing threshold.}

\item{save_dir}{Directory to save.}

\item{save_data}{Logical.}

\item{csv_filename}{Filename.}
}
\description{
Remove Missing Values from Stat Object or Data Frame
}
