% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{stat_normalize_process}
\alias{stat_normalize_process}
\title{Normalize Data in 'Stat' Object or Data Frame}
\usage{
stat_normalize_process(
  object,
  normalize_method = "log_transform",
  group_col = "group",
  max_unique_values = 5
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame containing the data to be normalized.
The 'Stat' object should contain a slot 'clean.data' for data and an optional slot
'group_col' for grouping information.}

\item{normalize_method}{A string specifying the normalization method. One of: "log_transform",
"min_max_scale", "z_score_standardize", "max_abs_scale", "center_data",
or "scale_data". Default is "log_transform".}

\item{group_col}{A string specifying the column that contains group information. This is used for
diagnostic purposes and is optional. Default is "group".}

\item{max_unique_values}{The maximum number of unique values to classify a column as categorical.
Default is 5. Used in \code{diagnose_variable_type} to classify variables.}
}
\value{
A 'Stat' object with updated 'scale.data' slot if the input is a 'Stat' object,
or a normalized data frame if the input is a data frame.
}
\description{
This function normalizes the numeric columns in the provided 'Stat' object or data frame using a selected method.
It checks for the existence of valid data and applies the chosen normalization method to the numeric columns.
The normalized data can either be returned as a data frame or integrated back into the 'Stat' object.
}
\examples{
# Example of using stat_normalize_process with a 'Stat' object
stat_obj <- Stat$new(clean.data = data, group_col = "group")
normalized_stat_obj <- stat_normalize_process(stat_obj, normalize_method = "z_score_standardize")
normalized_stat_obj@scale.data  # The normalized data

# Example of using stat_normalize_process with a data frame
data <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6), group = c("X", "Y", "X"))
normalized_data <- stat_normalize_process(data, normalize_method = "min_max_scale")
normalized_data  # The normalized data frame

}
