% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-2.R
\name{stat_onehot_encode}
\alias{stat_onehot_encode}
\title{Perform One-Hot Encoding on a Data Object}
\usage{
stat_onehot_encode(
  object,
  method = 1,
  group_col = "group",
  max_unique_values = 5
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame containing the data to be encoded.}

\item{method}{A method for encoding (not currently used, but kept for consistency with other functions).}

\item{group_col}{A string specifying the column that contains group information. This is optional, and if provided,
the group column will be moved to the end of the encoded data frame. Default is "group".}

\item{max_unique_values}{The maximum number of unique values a variable can have to be considered categorical.
Default is 5. This is used in \code{diagnose_variable_type} to determine which columns
should be one-hot encoded.}
}
\value{
The input \code{Stat} object with its data updated with one-hot encoded values in the 'clean.data' slot,
or a data frame containing the one-hot encoded data if the input was a data frame.
}
\description{
This function applies one-hot encoding to the categorical variables in the input object, which can either
be a \code{Stat} object or a data frame. It utilizes the \code{one_hot_encode} function to perform the encoding
and updates the \code{Stat} object with the encoded data, or returns the encoded data frame if the input is a data frame.
}
\examples{
# Example usage with a data frame
data <- data.frame(
  group = c("A", "B", "A", "B", "C"),
  category = c("X", "Y", "X", "Z", "Y")
)
encoded_data <- stat_onehot_encode(data, group_col = "group")
encoded_data

# Example usage with a 'Stat' object (assuming 'stat_object' is a valid Stat object)
stat_object <- stat_onehot_encode(stat_object, group_col = "group")

}
