% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_04_descriptive.R
\name{stat_onehot_encode}
\alias{stat_onehot_encode}
\title{Apply One-Hot Encoding to a Stat Object or Data Frame}
\usage{
stat_onehot_encode(
  object,
  method = 1,
  group_col = "group",
  max_unique_values = 5,
  save_dir = NULL,
  save_data = TRUE,
  csv_filename = "clean_data.csv"
)
}
\arguments{
\item{object}{An object of class \code{Stat} or a \code{data.frame}.}

\item{method}{Reserved for future use. Currently not used. Default is 1.}

\item{group_col}{A string specifying the name of the grouping column. This column will be
excluded from one-hot encoding and moved to the end of the result. Default is \code{"group"}.}

\item{max_unique_values}{The maximum number of unique values a variable can have to be
considered for one-hot encoding. Default is 5.}
}
\value{
A \code{Stat} object with updated one-hot encoded \code{clean.data} slot, or a
one-hot encoded \code{data.frame}.
}
\description{
This function applies one-hot encoding to categorical variables within a \code{Stat} object
or a regular \code{data.frame}, using \code{one_hot_encode()}. If a \code{Stat} object is provided,
the encoded result will be updated in its \code{clean.data} slot.
}
