% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_02_missing.R
\name{state_plot_missing_data}
\alias{state_plot_missing_data}
\title{Plot Missing Data for a Stat Object or Data Frame}
\usage{
state_plot_missing_data(
  object,
  palette_name = "Royal1",
  alpha = 0.9,
  save_plots = TRUE,
  save_dir = NULL,
  plot_width = 5,
  plot_height = 5,
  save_data = TRUE,
  var_filename = "var_missing_data.csv",
  sample_filename = "sample_missing_data.csv"
)
}
\arguments{
\item{object}{An object of class 'Stat' or a data frame. If it is a 'Stat' object, the function uses the
\code{raw.data} slot for missing data analysis. If it is a data frame, it directly performs the analysis.}

\item{palette_name}{A character string specifying the palette name for the plot colors. Default is \code{"Royal1"}.}

\item{alpha}{A numeric value between 0 and 1 specifying the transparency level of the density plots. Default is \code{0.9}.}

\item{save_plots}{A logical value indicating whether to save the plots as PNG files. Default is \code{TRUE}.}

\item{save_dir}{A character string specifying the directory where the plots will be saved. Default is \code{"here('StatObject', 'missing_info')"} (within the \code{StatObject} folder).}

\item{plot_width}{A numeric value specifying the width of the saved plots. Default is \code{5}.}

\item{plot_height}{A numeric value specifying the height of the saved plots. Default is \code{5}.}
}
\value{
If the input is a 'Stat' object, returns the updated 'Stat' object with the missing data information stored
in the \code{process.info} slot. If the input is a data frame, returns the missing data summary as a list.
}
\description{
This function generates and saves missing data plots (variable-wise and sample-wise distributions) for
an object of class 'Stat' or a data frame. It also updates the 'Stat' object with missing data information
and returns the object or the missing data summary, depending on the input type.
}
\examples{
\dontrun{
data("stat_obj_test")
# Generate missing data plots for a Stat object
updated_stat <- state_plot_missing_data(stat_obj_test, save_plots = TRUE)
# Generate missing data plots for a data frame
missing_info <- state_plot_missing_data(stat_obj_test@raw.data, save_plots = FALSE)
}
}
