% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_07_evaluate.R
\name{Sub_evaluation_results}
\alias{Sub_evaluation_results}
\title{Sub Evaluation Results}
\usage{
Sub_evaluation_results(object, seed = 123)
}
\arguments{
\item{object}{Subtyping object.}

\item{seed}{Seed.}
}
\value{
The updated \code{Subtyping} object (if input was Subtyping),
  or a named list of evaluation metrics.
}
\description{
Sub Evaluation Results
}
\examples{
\dontrun{
  # Assuming 'obj' is a Subtyping object with clustered.data
  obj <- Sub_evaluation_results(obj, seed = 123)
  print(obj@evaluation_results)
}
}
