% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_evaluate_clustering.R
\name{Sub_evaluation_results}
\alias{Sub_evaluation_results}
\title{Sub Evaluation Results}
\usage{
Sub_evaluation_results(object, seed = 123)
}
\arguments{
\item{object}{Subtyping object.}

\item{seed}{Seed.}
}
\description{
Sub Evaluation Results
}
