% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub3_kmeans.R
\name{Sub_kmeans_with_optimal_k}
\alias{Sub_kmeans_with_optimal_k}
\title{Sub K-means with Optimal K}
\usage{
Sub_kmeans_with_optimal_k(
  object,
  use_scaled_data = TRUE,
  palette_name = "Zissou1",
  save_plots = TRUE,
  save_dir = here("Subtyping", "cluster_results", "kmeans_result"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  seed = 123
)
}
\arguments{
\item{object}{Subtyping object.}

\item{use_scaled_data}{Logical.}

\item{palette_name}{Palette.}

\item{save_plots}{Logical.}

\item{save_dir}{Directory.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Base size.}

\item{seed}{Seed.}
}
\description{
Sub K-means with Optimal K
}
