% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub4_nmf.R
\name{Sub_nmf_estimate}
\alias{Sub_nmf_estimate}
\title{Sub NMF Estimate}
\usage{
Sub_nmf_estimate(
  object,
  rank_range = 2:4,
  seed = 8891,
  nrun = 10,
  method = "brunet",
  save_dir = here("Subtyping", "cluster_results", "nmf_results")
)
}
\arguments{
\item{object}{Subtyping object.}

\item{rank_range}{Rank range.}

\item{seed}{Seed.}

\item{nrun}{nrun.}

\item{method}{Method.}

\item{save_dir}{Directory.}
}
\description{
Sub NMF Estimate
}
