% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_04_nmf.R
\name{Sub_nmf_plot_consensus_all}
\alias{Sub_nmf_plot_consensus_all}
\title{Plot NMF consensus maps for all ranks}
\usage{
Sub_nmf_plot_consensus_all(object, save_dir = NULL, width = 12, height = 10)
}
\arguments{
\item{object}{Subtyping object.}

\item{save_dir}{Output directory.}
}
\description{
Plot NMF consensus maps for all ranks
}
