% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub2_normalize.R
\name{Sub_normalize_process}
\alias{Sub_normalize_process}
\title{Normalize Process for Subtyping}
\usage{
Sub_normalize_process(
  object,
  normalize_method = "min_max_scale",
  group_col = "group",
  max_unique_values = 5
)
}
\arguments{
\item{object}{Subtyping object or data frame.}

\item{normalize_method}{Normalization method.}

\item{group_col}{Group column.}

\item{max_unique_values}{Max unique values.}
}
\description{
Normalize Process for Subtyping
}
