% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyping_09_tsne.R
\name{Sub_plot_tsne}
\alias{Sub_plot_tsne}
\title{Sub Plot t-SNE}
\usage{
Sub_plot_tsne(
  object,
  palette_name = "AsteroidCity1",
  save_plots = TRUE,
  save_dir = file.path(get_output_dir("m3", "visualization_results"), "tsne"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  seed = 123
)
}
\arguments{
\item{object}{Subtyping object.}

\item{palette_name}{Palette.}

\item{save_plots}{Save plots.}

\item{save_dir}{Save dir.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Base size.}

\item{seed}{Seed.}
}
\description{
Sub Plot t-SNE
}
\examples{
\dontrun{
  # Assuming 'obj' is a Subtyping object with tsne results
  obj <- Sub_plot_tsne(obj, palette_name = "AsteroidCity1", save_plots = FALSE)
}
}
