% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub6_umap.R
\name{Sub_plot_umap}
\alias{Sub_plot_umap}
\title{Sub Plot UMAP}
\usage{
Sub_plot_umap(
  object,
  palette_name = "AsteroidCity1",
  save_plots = TRUE,
  save_dir = here("Subtyping", "visualization_results", "umap"),
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  seed = 123
)
}
\arguments{
\item{object}{Subtyping object.}

\item{palette_name}{Palette.}

\item{save_plots}{Save plots.}

\item{save_dir}{Save dir.}

\item{plot_width}{Width.}

\item{plot_height}{Height.}

\item{base_size}{Base size.}

\item{seed}{Seed.}
}
\description{
Sub Plot UMAP
}
