% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub5_tsne_vis.R
\name{Sub_tsne_analyse}
\alias{Sub_tsne_analyse}
\title{Sub t-SNE Analyse}
\usage{
Sub_tsne_analyse(
  object,
  perplexity = 30,
  dims = 2,
  iterations = 1000,
  use_scaled_data = TRUE
)
}
\arguments{
\item{object}{Subtyping object.}

\item{perplexity}{Perplexity.}

\item{dims}{Dims.}

\item{iterations}{Iterations.}

\item{use_scaled_data}{Logical.}
}
\description{
Sub t-SNE Analyse
}
