% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3_sub1_create_object.R
\docType{class}
\name{Subtyping-class}
\alias{Subtyping-class}
\alias{Subtyping}
\title{Subtyping Class}
\description{
Subtyping Class
}
\section{Slots}{

\describe{
\item{\code{clean.data}}{Data frame.}

\item{\code{info.data}}{Data frame.}

\item{\code{scale.data}}{Data frame.}

\item{\code{Optimal.cluster}}{Any.}

\item{\code{cluster.results}}{Any.}

\item{\code{visualization.results}}{Any.}

\item{\code{clustered.data}}{Data frame.}

\item{\code{evaluation_results}}{List.}
}}

