% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_02_framework.R
\name{surv_extract_task}
\alias{surv_extract_task}
\title{Extract Survival Task from PrognosiX or TaskSurv Object}
\usage{
surv_extract_task(object)
}
\arguments{
\item{object}{An object of class \code{PrognosiX} or \code{TaskSurv}.
If \code{PrognosiX}, the function extracts the survival data, time column,
and event column to create a \code{TaskSurv} object.}
}
\value{
A \code{\link[mlr3proba]{TaskSurv}} object.
}
\description{
This helper function extracts or returns a \code{TaskSurv} object from either
a \code{PrognosiX} S4 object or a \code{TaskSurv} object. It serves as a
unified interface for downstream functions that require a survival task.
}
\examples{
\dontrun{
library(mlr3proba)
library(survival)

# From TaskSurv
data("veteran", package = "survival")
task <- TaskSurv$new("veteran", backend = veteran, time = "time", event = "status")
task_out <- surv_extract_task(task)

# From PrognosiX object (if available)
data(pro_obj_test)
task_out <- surv_extract_task(pro_obj_test)
}
}
\seealso{
\code{\link{surv_create_surv_task}} for creating tasks from data frames
}
