% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_02_framework.R
\name{surv_filter_features_clinical}
\alias{surv_filter_features_clinical}
\title{Feature Selection Pipeline: Univariate to Lasso}
\usage{
surv_filter_features_clinical(object, p_threshold = 0.05)
}
\arguments{
\item{p_threshold}{P-value threshold for univariate filtering (default 0.05)}

\item{task}{TaskSurv object}
}
\value{
A list containing the filtered task and the univariate results table
}
\description{
Feature Selection Pipeline: Univariate to Lasso
}
