% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_02_framework.R
\name{surv_plot_time_dependent_auc}
\alias{surv_plot_time_dependent_auc}
\title{Plot Continuous Time-dependent AUC (Dynamic AUC)}
\usage{
surv_plot_time_dependent_auc(learner, object)
}
\arguments{
\item{learner}{Trained mlr3 learner}

\item{task}{TaskSurv object}
}
\value{
ggplot object
}
\description{
Plot Continuous Time-dependent AUC (Dynamic AUC)
}
