% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis_02_framework.R
\name{surv_predict_on_validation}
\alias{surv_predict_on_validation}
\title{Evaluate Model on a New Dataset (Validation/Test)}
\usage{
surv_predict_on_validation(learner, test_data, task_ref)
}
\arguments{
\item{learner}{Trained mlr3 learner}

\item{test_data}{New data frame (validation/test set)}

\item{task_ref}{Reference task to ensure column consistency}
}
\value{
Prediction object
}
\description{
Evaluate Model on a New Dataset (Validation/Test)
}
