% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_4_deg.R
\name{VarFeature_radarchart}
\alias{VarFeature_radarchart}
\title{Generate Variable Feature Radar Chart}
\usage{
VarFeature_radarchart(
  object,
  group_col = "group",
  palette_name = "Zissou1",
  plot_width = 5,
  plot_height = 5,
  save_dir = here("StatObject", "deg_info"),
  base_size = 14,
  title = "Variable Feature Radar Chart"
)
}
\arguments{
\item{object}{An object of class "Stat" or a data frame containing the variable features to plot.}

\item{group_col}{The name of the column representing the groups in the data (default: "group").}

\item{palette_name}{A string indicating the color palette for the plot .}

\item{plot_width}{The width of the plot (default: 5).}

\item{plot_height}{The height of the plot (default: 5).}

\item{save_dir}{The directory where the plot image will be saved (default: "here('StatObject', 'deg_info')").}

\item{base_size}{The base font size for the plot (default: 14).}

\item{title}{The title for the radar chart (default: "Variable Feature Radar Chart").}
}
\value{
If the input is a "Stat" object, the updated "Stat" object with the plot saved in \code{var.result}.
If the input is a data frame, it returns the generated radar chart plot.
}
\description{
This function generates a radar chart for variable features from either a "Stat" object or a data frame.
It uses the \code{plot_deg_radarchart} function to visualize the data and saves the plot as an image.
The chart is customized based on various input parameters such as the color palette, plot size, and title.
}
\examples{
# Example usage with a "Stat" object:
VarFeature_radarchart(stat_object, palette_name = "Zissou1", title = "Variable Feature Radar")

# Example usage with a data frame:
VarFeature_radarchart(df, palette_name = "Zissou1", title = "Variable Feature Radar")
}
