% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1_stat_4_deg.R
\name{VarFeature_ROC}
\alias{VarFeature_ROC}
\title{Plot ROC Curve for Variable Features in Statistical Object}
\usage{
VarFeature_ROC(
  object,
  group_col = "group",
  control = "health",
  case = "cancer",
  top_n = 5,
  palette_name = "Royal1",
  name_identity = "deg",
  data_type = "clean",
  save_dir = here("StatObject", "deg_info"),
  plot_width = 5,
  plot_height = 5,
  base_size = 10
)
}
\arguments{
\item{object}{An object of class \code{'Stat'} or a data frame containing the statistical test results.
If the input is a \code{Stat} object, it must contain the \code{last_test_sig} and data (either clean or scaled).}

\item{group_col}{The column in \code{data} or \code{object} representing the grouping variable (e.g., control vs case).
Default is \code{'group'}.}

\item{control}{The label used to indicate the control group in the data. Default is \code{'health'}.}

\item{case}{The label used to indicate the case group in the data. Default is \code{'cancer'}.}

\item{top_n}{The number of top features to consider based on log fold change. Default is 5.}

\item{palette_name}{The name of the palette to be used for the ROC curves.}

\item{name_identity}{The name or identifier used for the features in the \code{last_test_sig} data frame. Default is \code{'deg'}.}

\item{data_type}{The type of data to extract from the \code{Stat} object. Can be either \code{'clean'} or \code{'scale'}. Default is \code{'clean'}.}

\item{save_dir}{The directory path where the plot will be saved. Default is \code{here("StatObject", "deg_info")}.}

\item{plot_width}{The width of the saved plot in cm. Default is 5.}

\item{plot_height}{The height of the saved plot in cm. Default is 5.}

\item{base_size}{The base font size for the plot. Default is 14.}
}
\value{
The updated \code{Stat} object with the ROC plot stored in the \code{var.result} slot if the input is a \code{Stat} object.
If the input is a data frame, the ROC plot is returned directly.
}
\description{
This function generates a ROC curve for the top variable features identified in a statistical analysis.
It uses the results from a \code{Stat} object or a data frame, computes the ROC curves for the significant features,
and visualizes the AUC (Area Under the Curve) for each feature. The plot can be saved to a specified directory.
}
\examples{
VarFeature_ROC(object = stat_object)
VarFeature_ROC(object = df_data)
}
