% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_04_descriptive.R
\name{violin_plots}
\alias{violin_plots}
\title{Violin Plots for Numeric Variables}
\usage{
violin_plots(
  data,
  vars_per_plot = 1,
  save_dir = NULL,
  palette_name = "Zissou1",
  group_col = "group",
  max_unique_values = 5,
  sub_var = NULL,
  save_plots = TRUE,
  plot_width = 5,
  plot_height = 5,
  base_size = 14,
  stat_method = "wilcox.test",
  paired_comparison = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{vars_per_plot}{Integer. The number of variables (columns) to include
in each individual plot. Default is 1.}

\item{save_dir}{String. The directory to save the plots in. Default is `NULL`.
When `save_plots = TRUE` and this is `NULL`, a default figure directory is used.}

\item{palette_name}{String. The name of the color palette to use. Default `"Royal1"`.}

\item{group_col}{String. The name of the column in the data used to group
the samples. Default is `"group"`.}

\item{max_unique_values}{Integer. The maximum number of unique values
allowed for categorical variables. Defaults to 5.}

\item{sub_var}{Character vector. A subset of variable names (columns)
to include in the plot. If `NULL`, all numeric variables will be included.}

\item{save_plots}{Logical. If `TRUE`, the plots will be saved as PDF
files in the specified `save_dir`. Default is `TRUE`.}

\item{plot_width}{Numeric. The width of each saved plot (in inches). Default 5.}

\item{plot_height}{Numeric. The height of each saved plot (in inches). Default 5.}

\item{base_size}{Numeric. The base font size for the plot. Default 14.}

\item{stat_method}{Statistical method for comparisons.}

\item{paired_comparison}{Logical, whether to add pairwise comparisons.}
}
\value{
A list of ggplot objects containing the generated violin plots.
  If `save_plots` is `TRUE`, the plots are also saved as PDF files.
}
\description{
This function generates violin plots for numeric variables in the dataset,
with boxplots and jittered points for enhanced visualization. The plots
are grouped by a specified column and saved as PDF files, if desired.
It handles large numbers of variables by splitting them into multiple
plots, each containing a set of variables (defined by the `vars_per_plot`
argument). The function also allows for customization of plot appearance
and save locations.
}
\examples{
\dontrun{
vplots <- violin_plots(stat_obj_test@clean.data, group_col = "SWAB", save_plots = FALSE)
}
}
